### 2023-02-18 #5170 @editorialbot
[PRE REVIEW]: `SuperNOVA`: Semi-Parametric Identification and Estimation of Interaction and Effect Modification in Mixed Exposures using Stochastic Interventions in `R`
**Submitting author:** @blind-contours (<a href="http://orcid.org/0000-0002-5515-6307">David McCoy</a>)
**Repository:** https://github.com/blind-contours/SuperNOVA
**Branch with paper.md** (empty if default branch): main
**Version:** v1.0.0
