+++ sandbox/03_run_simulation.R
@@ -4,6 +4,7 @@ library(devtools)
 library(dplyr)
 library(SuperNOVA)
 library(magrittr)
+library(stringr)
 source(here("sandbox/02_fit_estimators.R"))
 
 # simulation parameters
