+++ sandbox/02_fit_estimators.R
@@ -35,10 +35,14 @@ fit_estimators <- function(data,
                            exposures,
                            outcome,
                            seed,
-                           effect_truth,
+                           true_effects,
+                           m14_effect_truth,
+                           m14_intxn_truth,
+                           true_em_effects,
                            deltas,
                            shift_var,
-                           cv_folds){
+                           cv_folds,
+                           var_sets){
   ## setup learners
 
   w <- data[, covars]
@@ -49,7 +53,6 @@ fit_estimators <- function(data,
     w = w,
     a = a,
     y = y,
-    var_sets = shift_var,
     delta = deltas,
     estimator = "tmle",
     fluctuation = "standard",
@@ -58,123 +61,96 @@ fit_estimators <- function(data,
     quantile_thresh = 0,
     verbose = TRUE,
     parallel = TRUE,
-    seed = seed
+    seed = seed,
+    var_sets = var_sets
   )
 
 
   indiv_shift_results <- sim_results$`Indiv Shift Results`
   em_results <- sim_results$`Effect Mod Results`
   joint_shift_results <- sim_results$`Joint Shift Results`
-  med_shift_results <- sim_results$`Mediation Shift Results`
 
-  if (is.null(indiv_shift_results[[shift_var]]) == FALSE) {
-    exposure_results <- indiv_shift_results[[shift_var]]
+  indiv_biases <- vector()
+  indiv_coverage <- vector()
+  indiv_estimates <- vector()
+
+  for (i in seq(true_effects)) {
+    true_effect <- true_effects[i]
+    exposure_results <- indiv_shift_results[[i]]
     exposure_results_pooled <- exposure_results[exposure_results$Fold == "Pooled TMLE", ]
     pooled_exposure_est <- exposure_results_pooled$Psi
-    pooled_bias <- effect - pooled_exposure_est
-    exposure_cov <- ifelse(
-      (exposure_results_pooled$`Lower CI` <= effect &
-        effect <= exposure_results_pooled$`Upper CI`), 1, 0
+    indiv_biases[[i]] <- true_effect - pooled_exposure_est
+    indiv_coverage[[i]] <- ifelse(
+      (exposure_results_pooled$`Lower CI` <= true_effect &
+         true_effect <= exposure_results_pooled$`Upper CI`), 1, 0
     )
-  } else {
-    pooled_exposure_est <- NULL
-    pooled_bias <- NULL
-    exposure_cov <- NULL
+
+    indiv_estimates[[i]] <- pooled_exposure_est
   }
 
+
   indiv_results <- list(
-    "pooled_exposure_est" = pooled_exposure_est,
-    "pooled_bias" = pooled_bias,
-    "exposure_cov" = exposure_cov
+    "indiv_est" = mean(indiv_estimates),
+    "indiv_bias" = mean(indiv_biases),
+    "indiv_cov" = mean(indiv_coverage)
   )
 
 
-  if (is.null(em_results$a_2w_2) == FALSE) {
[…]
