+++ sandbox/02_fit_estimators.R
@@ -1,9 +1,6 @@
 fit_estimators <- function(w,
                            a,
                            y,
-                           covars,
-                           exposures,
-                           outcome,
                            seed,
                            true_effects,
                            m14_effect_truth,
