+++ paper/paper.md
@@ -32,27 +32,25 @@ bibliography: paper.bib
 
 # Summary
 
-In environmental epidemiology studies, analysts are interested in understanding the joint impact of mixed exposures, i.e. a vector of exposures, on health outcomes while robustly adjusting for covariates. However, traditional statistical methods often make overly simplistic assumptions that lead to statistical quantities not being directly applicable to public policy decisions. As researchers, we are ultimately interested in causal questions such as how a decrease in exposures (like toxic chemicals) leads to a decrease in deleterious health outcomes (like cancer). For example, imagine a world where we made everyone decrease their dose of PFAS by 5%, what would the rate of thyroid cancer be? However, PFAS is a chemical class and many such chemicals are measured simultaneously in blood samples. How do we find shifts among the PFAS exposures that result in the greatest reduction in thyroid cancer rates? Furthermore, within the context of mixtures such as this, we are also interested in estimating statistical quantities of interaction and effect modification but doing so outside of a parametric context as nonlinear and non-additive relationships likely exist in mixed exposures. New methods in statistics are necessary to bridge this gap by delivering estimates on modified exposure policies. To address these limitations, the package SuperNOVA has been developed to use data-adaptive machine learning methods to identify the variables and variable sets that have the most explanatory power on an outcome of interest and applies non-parametric definitions of interaction and effect modification to these variable sets in a mixed exposure.
+Environmental epidemiology studies aim to understand the impact of mixed exposures on health outcomes while adjusting for covariates. However, traditional statistical methods make simplistic assumptions that may not be applicable to public policy decisions. Researchers are ultimately interested in answering causal questions, such as the impact of reducing toxic chemical exposures on adverse health outcomes like cancer. For example, in the case of PFAS, a class of chemicals measured simultaneously in blood samples, identifying the shifts that result in the greatest reduction in thyroid cancer rates can help more directly inform policy decisions on PFAS. In mixtures, nonlinear and non-additive relationships call for new statistical methods to estimate such modified exposure policies. To address these limitations, the open-source SuperNOVA package has been developed to use data-adaptive machine learning methods for identifying variable sets that have the most explanatory power on an outcome of interest. This package applies non-parametric definitions of interaction and effect modification to these variable sets in a mixed exposure, enabling researchers to explore modified treatment policies using stochastic interventions and answer causal questions.
 
 # Statement of Need
 
[…]
