+++ sandbox/NHANES/code/stash/nhanes_var_procesing.R
@@ -1,0 +1,518 @@
+################################################################################
+### Title:        NHANES Variable Processing
+### Author:        Aidan
+### Date Created:  4/14/2020
+################################################################################
+
+
+
+################################################################################
+### I. Environment, data loading
+################################################################################
+
+`%p%` <- function(x,y) paste0(x,y)
+library(tidyverse)
+library(magrittr)
+library(janitor)
+library(readxl)
+library(purrr)
+library(gtools)
+
+### get variable list
+var_list <-
+  read_excel("../input/Dictionary_1516.xlsx", sheet = "selected",
+             col_names = FALSE)
+
+names(var_list) <-
+  c("var_name", "var_description")
+
+
+### get data dictionary
+data_dic <-
+  read_excel("../input/Dictionary_1516.xlsx", sheet = "total",
+             col_names = TRUE)
+
+
+names(data_dic) <-
+  c("var_name", "var_description")
+
+
+### get relevant NHANES data
+rds_files <-
+  list.files("../input/",pattern="data.rds")
+
+
+empty_df <- c()
+
+
+### manually adding some dietary variables for now.
+diet_vars <- c("DR1TCAFF", "DR2TCAFF", "DRXTCAFF", 
+               "DR1TALCO", "DR2TACLO", "DRXTALCO",
+               "DR1TPROT", "DR2TPROT", "DRXTPROT",
+               "DR1TCHOL", "DR2TCHOL", "DRXTCHOL",
+               "DR1TSFAT", "DR2TSFAT", "DRXTSFAT",
+               "DR1TMFAT", "DR2TMFAT", "DRXTMFAT",
+               "DR1TPFAT", "DR2TPFAT", "DRXTPFAT")
+
+full_var_list <-
+  append(var_list$var_name, diet_vars)
+  
+
+for(paths in rds_files) {
+  
+
+  assign(gsub("_data.rds", "", paths),
+         
+         readRDS("../input/" %p% paths) %>%
+           select(matches(paste(full_var_list, collapse = "|"))) %>% 
+           mutate(name = gsub("_data.rds", "", paths))
+  )
+  
+  if(ncol(get(gsub("_data.rds", "", paths))) <= 2) {
+    empty_df <-
+      append(empty_df,
+             gsub("_data.rds", "", paths))}
+  
+  
+}
+
+### remove data if no variables of interest
+rm(list=empty_df)
+
+### NOTES:
+#### only the first BMX data has BMAUREXT and it is filled for 3 people.
+rm(BMX)
+
+
+
+################################################################################
+### II. Data Merging
+################################################################################
+
+
+### Get list of data
+data_list = list()
+j=1
+for(i in ls()) {
+  
+  if("data.frame" %in% class(get(i)) & !i %in% c("var_list", "data_dic")) {
+    data_list[[j]] = get(i) %>% mutate(name = i)
+    j=j+1
+  }
+}
+
+
+
+data_list <-
+  lapply(data_list,
+       function(x) x %>% 
+         mutate(year_begin = case_when(
+           grepl("_A", name) ~ 1999,
+           grepl("_B", name) ~ 2001,
+           grepl("_C", name) ~ 2003,
+           grepl("_D", name) ~ 2005,
+           grepl("_E", name) ~ 2007,
+           grepl("_F", name) ~ 2009,
+           grepl("_G", name) ~ 2011,
+           grepl("_H", name) ~ 2013,
+           grepl("_I", name) ~ 2015,
+           ### odd cases
+           name == "LAB06"   ~ 1999,
+           name == "LAB13AM" ~ 1999,
+           name == "DEMO"    ~ 1999,
+           name == "DRXTOT"  ~ 1999,
+           TRUE ~ NA_real_
+         )) %>% select(-name))
+
+
[…]
