+++ sandbox/NHANES/input/TEMP.txt
@@ -1,0 +1,1 @@
+This is a placeholder file for the input folder for PH 252 Final Project

[…]
