+++ sandbox/NHANES/code/stash/PFAS_mixtures2.Rmd
@@ -94,15 +94,19 @@ names(gluc_insulin_data_array) <- c('1999-2000', '2001-2002', '2003-2004',
                                     '2005-2006','2007-2008','2009-2010', 
                                     '2011-2012', '2013-2014', '2015-2016')
 
-df_list <- c(exercise_data_array)
+mbv_immunity_data_array <- c("L19_B.xpt", "L19_C.xpt")
+names(mbv_immunity_data_array) <- c('2001-2002', '2003-2004' )
 
-path <- '../input/'
+
+df_list <- c(mbv_immunity_data_array)
+
+path <- "/Users/davidmccoy/SuperNOVA/sandbox/NHANES/input/"
 ```
 
 ```{r generate_nhanes_functions}
 
-files <- list.files(path = "/Users/davidmccoy/Documents/PhD/Causal/final_proj/ph252d_group_project/input", 
-                    pattern = ".rds")
+# files <- list.files(path = "/Users/davidmccoy/Documents/PhD/Causal/final_proj/ph252d_group_project/input", 
+#                     pattern = ".rds")
 
 generate_nhanes_data <- function(data_array, save_path) {
   
@@ -131,7 +135,7 @@ generate_nhanes_data <- function(data_array, save_path) {
       }
   }
   return(output)
-} 
+
 ```
 
 ```{r run_gen_data}
