### 2023-02-23 7903857
update to earth learners - remove 3 way intxns
