+++ sandbox/NHANES/code/NIEHS_2019_workshop.R
@@ -25,24 +25,24 @@ exposures <- c("LBX074LA",
 
 NIEHS_2019 <- NIEHS_2019[complete.cases(NIEHS_2019[, exposures]), ]
 
-deltas <- list("LBX074LA" = 1,
-               "LBX099LA" = 1,
-               "LBX118LA" = 1,
-               "LBX138LA" = 1,
-               "LBX153LA" = 1,
-               "LBX170LA" = 1,
-               "LBX180LA" = 1,
-               "LBX187LA" = 1,
-               "LBX194LA" = 1,
-               "LBXD03LA" = 1,
-               "LBXD05LA" = 1,
-               "LBXD07LA" = 1,
-               "LBXF03LA" = 1,
-               "LBXF04LA" = 1,
-               "LBXF05LA" = 1,
-               "LBXF08LA" = 1,
-               "LBXHXCLA" = 1,
-               "LBXPCBLA" = 1)
+deltas <- list("LBX074LA" = 5,
+               "LBX099LA" = 5,
+               "LBX118LA" = 5,
+               "LBX138LA" = 5,
+               "LBX153LA" = 5,
+               "LBX170LA" = 5,
+               "LBX180LA" = 5,
+               "LBX187LA" = 5,
+               "LBX194LA" = 5,
+               "LBXD03LA" = 5,
+               "LBXD05LA" = 5,
+               "LBXD07LA" = 5,
+               "LBXF03LA" = 5,
+               "LBXF04LA" = 5,
+               "LBXF05LA" = 5,
+               "LBXF08LA" = 5,
+               "LBXHXCLA" = 5,
+               "LBXPCBLA" = 5)
 
 outcome <- "TELOMEAN"
 
@@ -70,7 +70,7 @@ nhanes_results <- SuperNOVA(
   a = a,
   y = y,
   delta = deltas,
-  n_folds = 10,
+  n_folds = 20,
   num_cores = 20,
   family = "continuous",
   quantile_thresh = 0,
