+++ sandbox/04_make_plots.R
@@ -1,25 +1,67 @@
 library(tidyverse)
 library(here)
 library(ggpubr)
+library(viridis)
+library(cividis)
+library(hrbrthemes)
 
-
-sim_results <- readRDS(
-  here("sandbox/data/SuperNOVA_sim.rds")
+sim_results_1 <- readRDS(
+  here("sandbox/data/SuperNOVA_1_sim.rds")
 )
+
+sim_results_2 <- readRDS(
+  here("sandbox/data/SuperNOVA_2_sim.rds")
+)
+
+sim_results_3 <- readRDS(
+  here("sandbox/data/SuperNOVA_3_sim.rds")
+)
+
+sim_results_4 <- readRDS(
+  here("sandbox/data/SuperNOVA_5_sim.rds")
+)
+
+sim_results <- rbind(sim_results_1, sim_results_2, sim_results_3, sim_results_4)
+
 
 sim_statistics <- sim_results %>%
   group_by(n_obs) %>%
   summarize(
-    indiv_bias = mean(indiv_bias),
-    indiv_est_sd = sd(indiv_bias),
-    indiv_est_MSE = est_bias^2 + est_sd^2,
-    indiv_CI_coverage = mean(indiv_cov)
+    mean_indiv_bias = mean(indiv_bias),
+    indiv_est_sd = sd(indiv_est),
+    indiv_est_MSE = mean_indiv_bias^2 + indiv_est_sd^2,
+    indiv_CI_coverage = mean(indiv_cov),
+    indiv_z = indiv_bias / indiv_est_sd,
+    mean_effect_mod_bias = mean(em_bias),
+    effect_mod_sd = sd(em_bias),
+    effect_est_MSE = mean_effect_mod_bias^2 + effect_mod_sd^2,
+    effect_mod_z = em_bias / effect_mod_sd,
+    em_cov = mean(em_cov),
+    mean_joint_bias = mean(joint_bias),
+    joint_est_sd =  sd(joint_est),
+    joint_est_MSE = mean_joint_bias^2 + joint_est_sd^2,
+    joint_cov = mean(joint_cov),
+    joint_z = joint_bias / joint_est_sd,
+    mean_intxn_bias = mean(intxn_bias),
+    intxn_est_sd = sd(intxn_est),
+    intxn_est_MSE = mean_intxn_bias^2 + intxn_est_sd^2,
+    intxn_cov = mean(intxn_cov),
+    intxn_z = intxn_bias/intxn_est_sd
   ) %>%
   mutate(
-    abs_bias = abs(est_bias),
-    sqrt_n_abs_bias = sqrt(n_obs) * abs(est_bias),
-    n_MSE = n_obs * est_MSE,
-  ) %>%
+    abs_indiv_bias = abs(mean_indiv_bias),
+    abs_effect_mod_bias = abs(mean_effect_mod_bias),
+    abs_joint_abs_bias =  abs(mean_joint_bias),
+    abs_intxn_bias = abs(mean_intxn_bias),
+    indiv_sqrt_n_abs_bias = sqrt(n_obs) * abs_indiv_bias,
+    effect_mod_sqrt_n_abs_bias = sqrt(n_obs) * abs_effect_mod_bias,
+    effect_mod_sqrt_n_abs_bias = sqrt(n_obs) * abs_joint_abs_bias,
+    effect_mod_sqrt_n_abs_bias = sqrt(n_obs) * abs_intxn_bias,
+    indiv_n_MSE = n_obs * indiv_est_MSE,
+    em_n_MSE = n_obs * effect_est_MSE,
+    joint_n_MSE = n_obs * joint_est_MSE,
+    intxn_n_MSE = n_obs * intxn_est_MSE
+    ) %>%
[…]
