### 2023-03-02 #5170 @blind-contours
@arfon - great thanks. I've rewritten the paper to be shorter.
