+++ sandbox/NHANES/code/NIEHS_2019_workshop.R
@@ -82,3 +82,8 @@ saveRDS(
   object = nhanes_results,
   file = here("sandbox/NHANES/output", paste0("SuperNOVA_", "nhanes_prime", ".rds"))
 )
+
+
+
+
+
