+++ sandbox/slurm_scripts/supernova_mediation_sim.sh
@@ -3,10 +3,10 @@
 #SBATCH --job-name=mediation_supernova
 #
 # Partition:
-#SBATCH --partition=savio3
+#SBATCH --partition=savio_bigmem
 #
 
-#SBATCH --qos=biostat_savio3_normal
+#SBATCH --qos=savio_lowprio
 #SBATCH --account=co_biostat
 
 # Number of nodes for use case:
