+++ NAMESPACE
@@ -38,6 +38,7 @@ export(simulate_mediation_data)
 export(tmle_exposhift)
 import(furrr)
 import(polspline)
+import(pracma)
 import(purrr)
 import(sl3)
 importFrom(MASS,mvrnorm)
@@ -49,6 +50,7 @@ importFrom(data.table,copy)
 importFrom(data.table,rbindlist)
 importFrom(data.table,set)
 importFrom(data.table,setnames)
+importFrom(dplyr,bind_rows)
 importFrom(dplyr,filter)
 importFrom(dplyr,group_by)
 importFrom(dplyr,mutate)
