### 2023-04-03 69210a3
stop on error in integration set to false, updates to sim
