+++ R/integrate_psi_aw_g_mc.R
@@ -1,4 +1,4 @@
-#' Integrate functions m and g
+#' Pseudo regression integration using monte carlo approaches
 #'
 #' @details Does the double integration as described in lemma 1
 #'
@@ -21,23 +21,22 @@
 #'  outcome mechanism at the natural value of the exposure Q(A, W) and an
 #'  upshift of the exposure Q(A + delta, W).
 
-integrate_m_g <- function(av, at, covars, w_names, q_model, g_model, exposure, g_delta, m_delta) {
+integrate_psi_aw_g_mc <- function(at, av, covars, w_names, pseudo_model, g_model, exposure, delta, psi_aw, n_samples = 1000) {
   at <- as.data.frame(at)
   av <- as.data.frame(av)
+  lower <- floor(min(min(av[exposure]), min(at[exposure])))
+  upper <- ceiling(max(max(av[exposure]), max(at[exposure])))
 
-  lower <- floor(min(av[exposure]))
-  upper <- ceiling(max(av[exposure]))
-
-  integrand <- function(a, row_data, covars, q_model, g_model, exposure, g_delta, m_delta, upper) {
+  integrand <- function(a, row_data, covars, pseudo_model, g_model, exposure, delta, upper) {
     row_data <- do.call("rbind", replicate(length(a), row_data, simplify = FALSE))
     new_data_m <- new_data_g <- row_data
-    new_data_m[exposure] <- ifelse(a + m_delta >= upper, upper, a + m_delta)
-    new_data_g[exposure] <- ifelse(a + g_delta >= upper, upper, a + g_delta)
+    new_data_m[exposure] <- a + delta #ifelse(a + delta >= upper, upper, a + delta)
+    new_data_g[exposure] <- a
 
     task_m <- sl3::sl3_Task$new(
       data = new_data_m,
       covariates = covars,
-      outcome = "y")
+      outcome = "pseudo_outcome")
 
     task_g <- sl3::sl3_Task$new(
       data = new_data_g,
@@ -45,11 +44,9 @@ integrate_m_g <- function(av, at, covars, w_names, q_model, g_model, exposure, g
       outcome = exposure,
     )
 
-    m_val <- q_model$predict(task_m)
-
+    m_val <- pseudo_model$predict(task_m)
     g_val <- g_model$predict(task_g)
     output <- m_val * g_val$likelihood
-
     return(output)
   }
 
@@ -57,16 +54,11 @@ integrate_m_g <- function(av, at, covars, w_names, q_model, g_model, exposure, g
 
   for (i in 1:nrow(av)) {
     row_data <- av[i, ]
+    sample_a <- runif(n_samples, lower, upper)
+    integral_values <- integrand(sample_a, row_data, covars, pseudo_model, g_model, exposure, delta, upper)
+    integral_result <- mean(integral_values) * (upper - lower)
 
-    integral_result <- stats::integrate(
-      function(a) integrand(a, row_data, covars, q_model, g_model, exposure, g_delta, m_delta, upper),
-      lower = lower,
-      upper = upper,
-      rel.tol = 0.0001,
-      subdivisions = 100,
-      stop.on.error = FALSE)$value
-
-    results[i] <- integral_result
+    results[i] <- psi_aw[i] - integral_result
   }
 
   return(results)
