### 2023-04-06 #5170 @csoneson
👋🏻 @wxwx1993, @stmcg - would you be interested in reviewing this submission for JOSS?

> SuperNOVA: Semi-Parametric Identification and Estimation of Interaction and Effect Modification in Mixed Exposures using Stochastic Interventions in R
