+++ R/indiv_stoch_shift_est_g_exp.R
@@ -35,12 +35,16 @@
 
 indiv_stoch_shift_est_g_exp <- function(exposure,
                                         delta,
-                                        pi_learner,
+                                        g_learner,
                                         covars,
                                         av,
                                         at,
                                         adaptive_delta,
-                                        hn_trunc_thresh) {
+                                        hn_trunc_thresh,
+                                        exposure_quantized,
+                                        lower_bound,
+                                        upper_bound,
+                                        outcome_type) {
   future::plan(future::sequential, gc = TRUE)
 
   # need a data set with the exposure stochastically shifted DOWNWARDS A-delta
@@ -49,7 +53,10 @@ indiv_stoch_shift_est_g_exp <- function(exposure,
   data.table::set(av_downshifted,
     j = exposure,
     value = shift_additive(
-      a = subset(av, select = exposure), delta = -delta
+      a = subset(av, select = exposure),
+      delta = -delta,
+      lower_bound = lower_bound,
+      upper_bound = upper_bound
     )
   )
 
@@ -58,7 +65,10 @@ indiv_stoch_shift_est_g_exp <- function(exposure,
   data.table::set(at_downshifted,
     j = exposure,
     value = shift_additive(
-      a = subset(at, select = exposure), delta = -delta
+      a = subset(at, select = exposure),
+      delta = -delta,
+      lower_bound = lower_bound,
+      upper_bound = upper_bound
     )
   )
 
@@ -67,7 +77,10 @@ indiv_stoch_shift_est_g_exp <- function(exposure,
   data.table::set(av_upshifted,
     j = exposure,
     value = shift_additive(
-      a = subset(av, select = exposure), delta = delta
+      a = subset(av, select = exposure),
+      delta = delta,
+      lower_bound = lower_bound,
+      upper_bound = upper_bound
     )
   )
 
@@ -75,7 +88,10 @@ indiv_stoch_shift_est_g_exp <- function(exposure,
   data.table::set(at_upshifted,
     j = exposure,
     value = shift_additive(
-      a = subset(at, select = exposure), delta = delta
+      a = subset(at, select = exposure),
+      delta = delta,
+      lower_bound = lower_bound,
+      upper_bound = upper_bound
     )
   )
 
@@ -84,7 +100,10 @@ indiv_stoch_shift_est_g_exp <- function(exposure,
   data.table::set(av_upupshifted,
     j = exposure,
     value = shift_additive(
-      a = subset(av, select = exposure), delta = 2 * delta
+      a = subset(av, select = exposure),
+      delta = 2 * delta,
+      lower_bound = lower_bound,
+      upper_bound = upper_bound
     )
   )
 
@@ -92,14 +111,18 @@ indiv_stoch_shift_est_g_exp <- function(exposure,
   data.table::set(at_upupshifted,
     j = exposure,
     value = shift_additive(
-      a = subset(at, select = exposure), delta = 2 * delta
+      a = subset(at, select = exposure),
+      delta = 2 * delta,
+      lower_bound = lower_bound,
+      upper_bound = upper_bound
     )
   )
 
   sl_task <- sl3::sl3_Task$new(
     data = at,
     outcome = exposure,
-    covariates = covars
+    covariates = covars,
+    outcome_type = outcome_type
   )
 
   sl_task_noshift_at <- sl3::sl3_Task$new(
@@ -153,7 +176,7 @@ indiv_stoch_shift_est_g_exp <- function(exposure,
     covariates = covars
   )
 
-  sl_fit <- pi_learner$train(sl_task)
[…]
