### 2023-04-09 68402e1
outcome type in sim med
