+++ sandbox/02_fit_estimators.R
@@ -146,7 +146,7 @@ fit_estimators_mediation <- function(w,
     estimator = "tmle",
     fluctuation = "standard",
     n_folds = 5,
-    family = "continuous",
+    outcome_type = "continuous",
     quantile_thresh = 0,
     verbose = TRUE,
     parallel = TRUE,
