### 2023-04-09 10534ae
speed glm issues
