### 2023-04-09 ad478ad
trying to remove speed glm
