+++ sandbox/mre.R
@@ -1,121 +1,0 @@
-library(here)
-library(sl3)
-library(tmle3)
-
-source(here("01_setup_data.R"))
-devtools::load_all(here())
-
-data <- sim_data(400)
-
-cv_folds <- 5
-
-## setup learners
-
-# continuous outcome learners
-fglm_contin_lrnr <- Lrnr_glm_fast$new()
-lasso_contin_lrnr <- Lrnr_glmnet$new(
-  alpha = 1, family = "gaussian",
-  nfolds = cv_folds
-)
-
-# binary outcome learners
-fglm_binary_lrnr <- Lrnr_glm_fast$new(family = binomial())
-lasso_binary_lrnr <- Lrnr_glmnet$new(
-  alpha = 1, family = "binomial",
-  nfolds = cv_folds
-)
-
-sl_contin <- Lrnr_sl$new(
-  learner = list(fglm_contin_lrnr, lasso_contin_lrnr),
-  metalearner = Lrnr_nnls$new()
-)
-
-sl_binary <- Lrnr_sl$new(
-  learner = list(fglm_binary_lrnr, lasso_binary_lrnr),
-  metalearner = Lrnr_nnls$new()
-)
-
-node_list <- list(
-  W = c("W1", "W2", "W3"),
-  A = "A",
-  Z = c("Z1", "Z2", "Z3"),
-  Y = "Y"
-)
-
-learner_list <- list(
-  Y = sl_contin,
-  A = sl_binary
-)
-
-tmle_spec_NDE_error <- tmle_NDE(
-  e_learners = sl_binary,
-  psi_Z_learners = sl_contin
-)
-
-tmle3(tmle_spec_NDE_error, data, node_list, learner_list)
-
-## Error in column_names[names(new_col_map)] <- new_col_map :
-##            'names' attribute [11] must be the same length as the vector [8]
-## Failed on chain
-## Error in self$compute_step() :
-##            Error in column_names[names(new_col_map)] <- new_col_map :
-##                       'names' attribute [11] must be the same length as the vector [8]
-
-sessionInfo()
-
-## R version 3.6.3 (2020-02-29)
-## Platform: x86_64-pc-linux-gnu (64-bit)
-## Running under: Ubuntu 18.04.4 LTS
-
-## Matrix products: default
-## BLAS:   /usr/lib/x86_64-linux-gnu/openblas/libblas.so.3
-## LAPACK: /usr/lib/x86_64-linux-gnu/libopenblasp-r0.2.20.so
-
-## locale:
-##  [1] LC_CTYPE=en_US.UTF-8       LC_NUMERIC=C
-##  [3] LC_TIME=en_US.UTF-8        LC_COLLATE=en_US.UTF-8
-##  [5] LC_MONETARY=en_US.UTF-8    LC_MESSAGES=en_US.UTF-8
-##  [7] LC_PAPER=en_US.UTF-8       LC_NAME=C
-##  [9] LC_ADDRESS=C               LC_TELEPHONE=C
-## [11] LC_MEASUREMENT=en_US.UTF-8 LC_IDENTIFICATION=C
-
-## attached base packages:
-## [1] stats     graphics  grDevices utils     datasets  methods   base
-
-## other attached packages:
-##  [1] tmle3mediate_0.0.1 testthat_2.3.2     forcats_0.4.0      stringr_1.4.0
-##  [5] dplyr_0.8.5        purrr_0.3.3        readr_1.3.1        tidyr_1.0.2
-##  [9] tibble_3.0.1       ggplot2_3.3.0      tidyverse_1.3.0    tmle3_0.1.7
-## [13] sl3_1.3.7          here_0.1
-
-## loaded via a namespace (and not attached):
[…]
