+++ R/indiv_stoch_shift_est_g_exp.R
@@ -44,7 +44,8 @@ indiv_stoch_shift_est_g_exp <- function(exposure,
                                         exposure_quantized,
                                         lower_bound,
                                         upper_bound,
-                                        outcome_type) {
+                                        outcome_type,
+                                        density_type) {
   future::plan(future::sequential, gc = TRUE)
 
   # need a data set with the exposure stochastically shifted DOWNWARDS A-delta
@@ -118,96 +119,128 @@ indiv_stoch_shift_est_g_exp <- function(exposure,
     )
   )
 
-  sl_task <- sl3::sl3_Task$new(
-    data = at,
-    outcome = exposure,
-    covariates = covars,
-    outcome_type = outcome_type
-  )
+  if (density_type == "sl") {
+    sl_task <- sl3::sl3_Task$new(
+      data = at,
+      outcome = exposure,
+      covariates = covars,
+      outcome_type = outcome_type
+    )
 
-  sl_task_noshift_at <- sl3::sl3_Task$new(
-    data = at,
-    outcome = exposure,
-    covariates = covars
-  )
+    sl_task_noshift_at <- sl3::sl3_Task$new(
+      data = at,
+      outcome = exposure,
+      covariates = covars
+    )
 
-  sl_task_noshift_av <- sl3::sl3_Task$new(
-    data = av,
-    outcome = exposure,
-    covariates = covars
-  )
+    sl_task_noshift_av <- sl3::sl3_Task$new(
+      data = av,
+      outcome = exposure,
+      covariates = covars
+    )
 
-  # sl3 task for data with exposure shifted DOWNWARDS A-delta
-  sl_task_downshifted_at <- sl3::sl3_Task$new(
-    data = at_downshifted,
-    outcome = exposure,
-    covariates = covars
-  )
+    # sl3 task for data with exposure shifted DOWNWARDS A-delta
+    sl_task_downshifted_at <- sl3::sl3_Task$new(
+      data = at_downshifted,
+      outcome = exposure,
+      covariates = covars
+    )
 
-  sl_task_downshifted_av <- sl3::sl3_Task$new(
-    data = av_downshifted,
-    outcome = exposure,
-    covariates = covars
-  )
+    sl_task_downshifted_av <- sl3::sl3_Task$new(
+      data = av_downshifted,
+      outcome = exposure,
+      covariates = covars
+    )
 
-  # sl3 task for data with exposure shifted UPWARDS A+delta
-  sl_task_upshifted_at <- sl3::sl3_Task$new(
-    data = at_upshifted,
-    outcome = exposure,
-    covariates = covars
-  )
+    # sl3 task for data with exposure shifted UPWARDS A+delta
+    sl_task_upshifted_at <- sl3::sl3_Task$new(
+      data = at_upshifted,
+      outcome = exposure,
+      covariates = covars
+    )
 
-  sl_task_upshifted_av <- sl3::sl3_Task$new(
-    data = av_upshifted,
-    outcome = exposure,
-    covariates = covars
-  )
+    sl_task_upshifted_av <- sl3::sl3_Task$new(
+      data = av_upshifted,
+      outcome = exposure,
+      covariates = covars
+    )
 
-  # sl3 task for data with exposure shifted UPWARDS A+2delta
-  sl_task_upupshifted_at <- sl3::sl3_Task$new(
-    data = at_upupshifted,
-    outcome = exposure,
-    covariates = covars
-  )
[…]
