+++ R/integrate_psi_g_mc.R
@@ -26,22 +26,22 @@
 #' @return A \code{data.table} with two columns, containing estimates of the
 #'  outcome mechanism at the natural value of the exposure Q(A, W) and an
 #'  upshift of the exposure Q(A + delta, W)
-integrate_psi_g_mc <- function(av, at, covars, w_names, q_model, r_model, g_model, exposure, mediator, delta, n_samples, n_iterations) {
+integrate_psi_g_mc <- function(av, at, covars, w_names, q_model, r_model, g_model, exposure, mediator, delta, n_samples, n_iterations, density_type) {
   av <- as.data.frame(av)
   at <- as.data.frame(at)
 
   lower_z <- floor(min(min(av[[mediator]]), min(at[[mediator]])))
   upper_z <- ceiling(max(max(av[[mediator]]), max(at[[mediator]])))
 
-  lower_a <- floor(min(min(av[[exposure]]), min(at[[exposure]])))
-  upper_a <- ceiling(max(max(av[[exposure]]), max(at[[exposure]])))
+  lower_a <- min(min(av[[exposure]]), min(at[[exposure]]))
+  upper_a <- max(max(av[[exposure]]), max(at[[exposure]]))
 
-  integrand_m_r <- function(sample_z_inner, row_data, covars, w_names, q_model, r_model, exposure, mediator, delta, upper_a) {
+  integrand_m_r <- function(sample_z_inner, row_data, covars, w_names, q_model, r_model, exposure, mediator, delta, upper_a, density_type) {
     row_data <- do.call("rbind", replicate(length(sample_z_inner), row_data, simplify = FALSE))
     new_data_m <- new_data_r <- row_data
 
     new_data_m[mediator] <- sample_z_inner
-    new_data_m[exposure] <- new_data_m[[exposure]] + delta # ifelse(new_data_m[[exposure]] + delta >= upper_a, upper_a, new_data_m[[exposure]] + delta)
+    new_data_m[exposure] <- ifelse(new_data_m[[exposure]] + delta >= upper_a, upper_a, new_data_m[[exposure]] + delta)
 
     new_data_r[mediator] <- sample_z_inner
 
@@ -69,7 +69,7 @@ integrate_psi_g_mc <- function(av, at, covars, w_names, q_model, r_model, g_mode
     row_data <- do.call("rbind", replicate(n_samples, row_data, simplify = FALSE))
     new_data_m <- new_data_g <- new_data_r <- row_data
 
-    new_data_m[exposure] <- sample_a + delta # ifelse(sample_a + delta >= upper_a, upper_a, sample_a + delta)
+    new_data_m[exposure] <- ifelse(sample_a + delta >= upper_a, upper_a, sample_a + delta)
     new_data_m[mediator] <- sample_z
 
     new_data_g[exposure] <- sample_a
@@ -81,12 +81,6 @@ integrate_psi_g_mc <- function(av, at, covars, w_names, q_model, r_model, g_mode
       outcome = "y"
     )
 
-    task_g <- sl3::sl3_Task$new(
-      data = new_data_g,
-      covariates = c(w_names),
-      outcome = exposure,
-    )
-
     task_r <- sl3::sl3_Task$new(
       data = new_data_r,
       covariates = c(w_names),
@@ -94,10 +88,22 @@ integrate_psi_g_mc <- function(av, at, covars, w_names, q_model, r_model, g_mode
     )
 
     m_val <- q_model$predict(task_m)
-    g_val <- g_model$predict(task_g)
     r_val <- r_model$predict(task_r)
 
-    output <- m_val * g_val$likelihood * r_val$likelihood
+    if (density_type == "sl") {
+      task_g <- sl3::sl3_Task$new(
+        data = new_data_g,
[…]
