+++ sandbox/02_fit_estimators.R
@@ -145,7 +145,7 @@ fit_estimators_mediation <- function(w,
     deltas = deltas,
     estimator = "tmle",
     fluctuation = "standard",
-    n_folds = 5,
+    n_folds = 2,
     outcome_type = "continuous",
     quantile_thresh = 0,
     verbose = TRUE,
