+++ NAMESPACE
@@ -36,7 +36,6 @@ export(joint_stoch_shift_est_Q)
 export(joint_stoch_shift_est_g_exp)
 export(joint_stoch_shift_est_z_exp)
 export(list_rules_party)
-export(make_density_superlearner)
 export(scale_to_original)
 export(scale_to_unit)
 export(simulate_data)

[…]
