+++ R/fit_basis_estimators.R
@@ -74,7 +74,7 @@ fit_basis_estimators <- function(at,
       exposure_drivers <- exposure_drivers[!is.na(exposure_drivers)]
       mediator_exposures[[mediator]] <- unique(unlist(exposure_drivers))
     }
-  }else{
+  } else {
     mediator_exposures <- NULL
   }
 

[…]
