### 2023-04-12 0b96c0c
remove print update sims
