### 2023-04-13 2716693
truncate the a/e in pseudo, scale to keep model consistent then backscale.
