+++ sandbox/04_make_plots.R
@@ -6,107 +6,86 @@ library(cividis)
 library(hrbrthemes)
 
 sim_results_1 <- readRDS(
-  here("sandbox/data/SuperNOVA_1_sim.rds")
+  here("sandbox/data/SuperNOVA_mediation.rds")
 )
 
-sim_results_2 <- readRDS(
-  here("sandbox/data/SuperNOVA_2_sim.rds")
-)
+# sim_results_2 <- readRDS(
+#   here("sandbox/data/SuperNOVA_2_sim.rds")
+# )
+#
+# sim_results_3 <- readRDS(
+#   here("sandbox/data/SuperNOVA_3_sim.rds")
+# )
+#
+# sim_results_4 <- readRDS(
+#   here("sandbox/data/SuperNOVA_5_sim.rds")
+# )
+#
+# sim_results_5 <- readRDS(
+#   here("sandbox/data/SuperNOVA_6_sim.rds")
+# )
+#
+# sim_results_6 <- readRDS(
+#   here("sandbox/data/SuperNOVA_7_sim.rds")
+# )
 
-sim_results_3 <- readRDS(
-  here("sandbox/data/SuperNOVA_3_sim.rds")
-)
 
-sim_results_4 <- readRDS(
-  here("sandbox/data/SuperNOVA_5_sim.rds")
-)
-
-sim_results_5 <- readRDS(
-  here("sandbox/data/SuperNOVA_6_sim.rds")
-)
-
-sim_results_6 <- readRDS(
-  here("sandbox/data/SuperNOVA_7_sim.rds")
-)
-
-
-sim_results <- rbind(sim_results_1, sim_results_2, sim_results_3, sim_results_4, sim_results_5, sim_results_6)
+sim_results <- rbind(sim_results_1)
 
 
 sim_statistics <- sim_results %>%
   group_by(n_obs) %>%
   summarize(
-    mean_indiv_bias = mean(indiv_bias),
-    indiv_est_sd = sd(indiv_est),
-    indiv_est_MSE = mean_indiv_bias^2 + indiv_est_sd^2,
-    indiv_CI_coverage = mean(indiv_cov),
-    indiv_z = indiv_bias / indiv_est_sd,
-    mean_effect_mod_bias = mean(em_bias),
-    effect_mod_sd = sd(em_bias),
-    effect_est_MSE = mean_effect_mod_bias^2 + effect_mod_sd^2,
-    effect_mod_z = em_bias / effect_mod_sd,
-    em_cov = mean(em_cov),
-    mean_joint_bias = mean(joint_bias),
-    joint_est_sd =  sd(joint_est),
-    joint_est_MSE = mean_joint_bias^2 + joint_est_sd^2,
-    joint_cov = mean(joint_cov),
-    joint_z = joint_bias / joint_est_sd,
-    mean_intxn_bias = mean(intxn_bias),
-    intxn_est_sd = sd(intxn_est),
-    intxn_est_MSE = mean_intxn_bias^2 + intxn_est_sd^2,
-    intxn_cov = mean(intxn_cov),
-    intxn_z = intxn_bias/intxn_est_sd
+    mean_nde_bias_pseudo = mean(pseudo_reg_nde_bias),
+    nde_est_sd_pseudo = sd(pseudo_reg_nde_est),
+    nde_est_MSE_pseudo = mean_nde_bias_pseudo^2 + nde_est_sd_pseudo^2,
+    nde_pseudo_CI_coverage = mean(pseudo_reg_nde_cov),
+    #
+    mean_nde_bias_int = mean(integrated_nde_bias),
+    nde_est_sd_int = sd(integrated_nde_est),
[…]
