### 2023-04-15 #5170 @csoneson
👋🏻 @simonpcouch, @mhu48 - would you be interested in reviewing this submission for JOSS?

> SuperNOVA: Semi-Parametric Identification and Estimation of Interaction and Effect Modification in Mixed Exposures using Stochastic Interventions in R
