+++ sandbox/03_run_med_simulation_cont.R
@@ -31,9 +31,9 @@ nie_a1_quant <- full_data$nie_a1_quant
 ate_a1_quant <- full_data$ate_a1_quant
 
 covars <- c("w_1", "w_2", "w_3", "w_4", "w_5")
-exposures <- c("a_1_quant")
-mediators <- c("z_1_quant")
-outcome <- "y_quant"
+exposures <- c("a_1")
+mediators <- c("z_1")
+outcome <- "y"
 
 # perform simulation across sample sizes
 sim_results_df <- data.frame()
@@ -63,9 +63,9 @@ for (sample_size in n_obs) {
       z = z,
       y = y,
       seed = seed,
-      nde_effects = c(nde_a1_quant),
-      nie_effects = c(nie_a1_quant),
-      ate_effects = c(ate_a1_quant),
+      nde_effects = c(nde_a1),
+      nie_effects = c(nie_a1),
+      ate_effects = c(ate_a1),
       deltas = list("a" = 1),
       cv_folds = n_fold,
       num_cores = n_core,
@@ -85,9 +85,9 @@ for (sample_size in n_obs) {
       z = z,
       y = y,
       seed = seed,
-      nde_effects = c(nde_a1_quant),
-      nie_effects = c(nie_a1_quant),
-      ate_effects = c(ate_a1_quant),
+      nde_effects = c(nde_a1),
+      nie_effects = c(nie_a1),
+      ate_effects = c(ate_a1),
       deltas = list("a" = 1),
       cv_folds = n_fold,
       num_cores = n_core,
@@ -114,5 +114,5 @@ for (sample_size in n_obs) {
 # save results to file
 saveRDS(
   object = sim_results_df,
-  file = here("sandbox/data", paste0("SuperNOVA_", "mediation_quant", ".rds"))
+  file = here("sandbox/data", paste0("SuperNOVA_", "mediation_cont", ".rds"))
 )
