+++ R/indiv_stoch_shift_est_g_exp.R
@@ -1,24 +1,51 @@
-#' Estimate the Exposure Mechanism via Generalized Propensity Score for One
-#' Exposure Variable
+#' Estimate the Exposure Mechanism via Generalized Propensity Score for One Exposure Variable
 #'
-#' @details Compute the propensity score (exposure mechanism) for the observed
-#'  data, including the shift. This gives the propensity score for the observed
-#'  data (at the observed A) the counterfactual shifted exposure levels (at
-#'  {A - delta}, {A + delta}, and {A + 2 * delta}).
+#' @details This function computes the generalized propensity score (GPS) for the observed
+#' data with one exposure variable, considering different shift levels. It estimates the GPS at
+#' the observed data (at the observed A), and at the counterfactual shifted exposure levels
+#' (at {A - delta}, {A + delta}, and {A + 2 * delta}).
 #'
-#' @param exposure A \code{character} labeling the exposure variable.
-#' @param covars A \code{character} labeling the covariate variables
-#' @param delta A \code{numeric} value identifying a shift in the observed
-#'  value of the exposure under which observations are to be evaluated.
-#' @param pi_learner Object containing a set of instantiated learners
-#'  from \pkg{sl3}, to be used in fitting an ensemble model.
-#' @param av A \code{dataframe} of validation data specific to the fold
-#' @param at A \code{dataframe} of training data specific to the fold
-#' @param adaptive_delta Whether or not to adaptively adjust delta based on
-#' positivity (estimated from the clever covariate) meeting the hn_trunc_thresh
-#' level
-#' @param hn_trunc_thresh Level of the clever covariate in the adaptive delta
-#' procedure
+#' @param exposure A \code{character} representing the label of the exposure variable. This
+#' variable should be a column name in the input data.
+#' @param covars A \code{character} vector representing the labels of the covariate variables.
+#' These variables should be column names in the input data and serve as control variables in
+#' the GPS estimation.
+#' @param delta A \code{numeric} value specifying the shift in the observed value of the
+#' exposure for evaluating counterfactual observations. Positive values will result in upward
+#' shifts, while negative values will result in downward shifts.
+#' @param pi_learner An object containing a set of instantiated learners from the \pkg{sl3}
+#' package, to be used in fitting an ensemble model for GPS estimation. Learners should be
+#' chosen based on the structure of the data and the relationships between exposure, covariates,
+#' and outcome.
+#' @param av A \code{dataframe} containing validation data specific to the fold. This data is
+#' used to evaluate the performance of the GPS model during cross-validation.
+#' @param at A \code{dataframe} containing training data specific to the fold. This data is
+#' used to fit the GPS model during cross-validation.
+#' @param adaptive_delta A \code{logical} indicating whether to adaptively adjust delta based on
+#' positivity (estimated from the clever covariate) meeting the hn_trunc_thresh level. If
+#' \code{TRUE}, the function will adjust delta to ensure sufficient overlap in the GPS
+#' distributions.
+#' @param hn_trunc_thresh A \code{numeric} value specifying the level of the clever covariate
+#' in the adaptive delta procedure. It represents the minimum proportion of observations in each
+#' exposure group to be included when adjusting delta.
+#' @param exposure_quantized A \code{logical} indicating whether the exposure variable has been
+#' discretized. If \code{TRUE}, multinomial learners will be used instead of density learners
+#' for the GPS estimation.
[…]
