+++ R/integrate_psi_aw_g.R
@@ -21,7 +21,7 @@
 #'  outcome mechanism at the natural value of the exposure Q(A, W) and an
 #'  upshift of the exposure Q(A + delta, W).
 
-integrate_psi_aw_g <- function(at, av, covars, w_names, pseudo_model, g_model, exposure, delta, psi_aw) {
+integrate_psi_aw_g <- function(at, av, covars, w_names, pseudo_model, g_model, exposure, delta, psi_aw, integration_method = "AQ") {
   at <- as.data.frame(at)
   av <- as.data.frame(av)
   lower <- min(av[exposure])
@@ -56,14 +56,21 @@ integrate_psi_aw_g <- function(at, av, covars, w_names, pseudo_model, g_model, e
 
   for (i in 1:nrow(av)) {
     row_data <- av[i, ]
-    integral_result <- stats::integrate(
-      function(a) integrand(a, row_data, covars, pseudo_model, g_model, exposure, delta, upper),
-      lower = lower,
-      upper = upper,
-      rel.tol = 0.0001,
-      subdivisions = 1000,
-      stop.on.error = FALSE
-    )$value
+
+    if (integration_method == "MC") {
+      sample_a <- runif(n_samples, lower, upper)
+      integral_values <- integrand(sample_a, row_data, covars, pseudo_model, g_model, exposure, delta, upper)
+      integral_result <- mean(integral_values) * (max(sample_a) - min(sample_a))
+    } else if (integration_method == "AQ") {
+      integral_result <- stats::integrate(
+        function(a) integrand(a, row_data, covars, pseudo_model, g_model, exposure, delta, upper),
+        lower = lower,
+        upper = upper,
+        rel.tol = 0.001,
+        subdivisions = 100,
+        stop.on.error = FALSE
+      )$value
+    }
 
     results[i] <- psi_aw[i] - integral_result
   }
