+++ sandbox/02_fit_estimators.R
@@ -137,6 +137,7 @@ fit_estimators_mediation <- function(w,
                                      num_cores,
                                      var_sets,
                                      exposure_quantized,
+                                     mediator_quantized,
                                      n_mc_sample,
                                      density_type,
                                      integration_method = "MC"){
@@ -158,6 +159,7 @@ fit_estimators_mediation <- function(w,
     var_sets = var_sets,
     num_cores = num_cores,
     exposure_quantized = exposure_quantized,
+    mediator_quantized = mediator_quantized,
     density_type = density_type,
     n_mc_sample = n_mc_sample,
     integration_method = integration_method
