### 2023-04-18 670b077
Just do quant in 03_run med sims
