+++ man/indiv_stoch_shift_est_g_exp.Rd
@@ -2,8 +2,7 @@
 % Please edit documentation in R/indiv_stoch_shift_est_g_exp.R
 \name{indiv_stoch_shift_est_g_exp}
 \alias{indiv_stoch_shift_est_g_exp}
-\title{Estimate the Exposure Mechanism via Generalized Propensity Score for One
-Exposure Variable}
+\title{Estimate the Exposure Mechanism via Generalized Propensity Score for One Exposure Variable}
 \usage{
 indiv_stoch_shift_est_g_exp(
   exposure,
@@ -24,40 +23,74 @@ indiv_stoch_shift_est_g_exp(
 )
 }
 \arguments{
-\item{exposure}{A \code{character} labeling the exposure variable.}
+\item{exposure}{A \code{character} representing the label of the exposure variable. This
+variable should be a column name in the input data.}
 
-\item{delta}{A \code{numeric} value identifying a shift in the observed
-value of the exposure under which observations are to be evaluated.}
+\item{delta}{A \code{numeric} value specifying the shift in the observed value of the
+exposure for evaluating counterfactual observations. Positive values will result in upward
+shifts, while negative values will result in downward shifts.}
 
-\item{covars}{A \code{character} labeling the covariate variables}
+\item{covars}{A \code{character} vector representing the labels of the covariate variables.
+These variables should be column names in the input data and serve as control variables in
+the GPS estimation.}
 
-\item{av}{A \code{dataframe} of validation data specific to the fold}
+\item{av}{A \code{dataframe} containing validation data specific to the fold. This data is
+used to evaluate the performance of the GPS model during cross-validation.}
 
-\item{at}{A \code{dataframe} of training data specific to the fold}
+\item{at}{A \code{dataframe} containing training data specific to the fold. This data is
+used to fit the GPS model during cross-validation.}
 
-\item{adaptive_delta}{Whether or not to adaptively adjust delta based on
-positivity (estimated from the clever covariate) meeting the hn_trunc_thresh
-level}
+\item{adaptive_delta}{A \code{logical} indicating whether to adaptively adjust delta based on
+positivity (estimated from the clever covariate) meeting the hn_trunc_thresh level. If
+\code{TRUE}, the function will adjust delta to ensure sufficient overlap in the GPS
+distributions.}
 
-\item{hn_trunc_thresh}{Level of the clever covariate in the adaptive delta
-procedure}
+\item{hn_trunc_thresh}{A \code{numeric} value specifying the level of the clever covariate
+in the adaptive delta procedure. It represents the minimum proportion of observations in each
+exposure group to be included when adjusting delta.}
 
-\item{pi_learner}{Object containing a set of instantiated learners
-from \pkg{sl3}, to be used in fitting an ensemble model.}
+\item{exposure_quantized}{A \code{logical} indicating whether the exposure variable has been
+discretized. If \code{TRUE}, multinomial learners will be used instead of density learners
+for the GPS estimation.}
+
+\item{lower_bound}{A \code{numeric} value specifying the lower bound of the exposure variable
+to prevent shifting past this limit during the GPS estimation.}
+
+\item{upper_bound}{A \code{numeric} value specifying the upper bound of the exposure variable
+to prevent shifting past this limit during the GPS estimation.}
+
+\item{outcome_type}{A \code{character} specifying whether the outcome is 'categorical' or
+'continuous' based on the discretization of the exposure variable. This information is used
+to determine the appropriate learner type for the GPS model.}
+
+\item{density_type}{A \code{character} specifying the type of density estimator to use for
+GPS estimation. Possible options are 'SL' (Super Learner) or 'HAL' (Highly Adaptive Lasso).}
+
[…]
