### 2023-04-18 7654175
udpates to Q and g_delta.
