### 2023-04-18 3ed41a1
updated the sim scripts for med quant and exp quant
