+++ sandbox/03_run_med_simulation_quant.R
@@ -9,13 +9,13 @@ source(here("sandbox/simulate_discrete_data.R"))
 load_all()
 
 # simulation parameters
-n_sim <- 2 # number of simulations
+n_sim <- 5 # number of simulations
 n_obs <- c(250, 500, 1000, 1500, 2500, 3000)
 p0_obs <- 100000
 n_mc_sample <- 1000
 
-n_core <- 5
-n_fold <- 4
+n_core <- 20
+n_fold <- 10
 
 # Generate simulated data -----------------
 
