+++ R/integrate_psi_aw_g_quant.R
@@ -33,7 +33,7 @@ integrate_psi_aw_g_quant <- function(at, av, covars, w_names, pseudo_model, g_mo
     task_m <- sl3::sl3_Task$new(
       data = new_data_m,
       covariates = covars,
-      outcome = "pseudo_outcome_scaled"
+      outcome = "pseudo_outcome"
     )
 
     task_g <- sl3::sl3_Task$new(
