+++ R/integrate_psi_g.R
@@ -26,17 +26,17 @@
 #'  outcome mechanism at the natural value of the exposure Q(A, W) and an
 #'  upshift of the exposure Q(A + delta, W).
 
-integrate_psi_g <- function(data, covars, w_names, q_model, r_model, g_model, exposure, mediator, delta, integration_method, n_samples) {
-  data <- as.data.frame(data)
+integrate_psi_g <- function(at, av, covars, w_names, q_model, r_model, g_model, exposure, mediator, delta, integration_method, n_samples, density_type) {
+  av <- as.data.frame(av)
 
-  lower_z <- min(data[[mediator]])
-  upper_z <- max(data[[mediator]])
+  lower_z <- min(av[[mediator]])
+  upper_z <- max(av[[mediator]])
 
-  lower_a <- min(data[[exposure]])
-  upper_a <- max(data[[exposure]])
+  lower_a <- min(av[[exposure]])
+  upper_a <- max(av[[exposure]])
 
 
-  integrand_m_r <- function(z, row_data, covars, w_names, q_model, r_model, exposure, mediator, delta, upper_a) {
+  integrand_m_r <- function(z, row_data, covars, w_names, q_model, r_model, exposure, mediator, delta, upper_a, density_type) {
     row_data <- do.call("rbind", replicate(length(z), row_data, simplify = FALSE))
     new_data_m <- new_data_r <- row_data
 
@@ -51,21 +51,27 @@ integrate_psi_g <- function(data, covars, w_names, q_model, r_model, g_model, ex
       outcome = "y"
     )
 
-    task_r <- sl3::sl3_Task$new(
-      data = new_data_r,
-      covariates = c(w_names),
-      outcome = mediator,
-    )
-
     m_val <- q_model$predict(task_m)
-    r_val <- r_model$predict(task_r)
 
-    output <- m_val * r_val$likelihood
+    if (density_type == "sl") {
+      task_r <- sl3::sl3_Task$new(
+        data = new_data_r,
+        covariates = c(w_names),
+        outcome = mediator,
+      )
+
+      r_val <- r_model$predict(task_r)
+      output <- m_val * r_val$likelihood
+    } else {
+      r_val <- suppressMessages(predict(r_model, new_A = new_data_r[[mediator]], new_W = new_data_r[w_names]))
+      output <- m_val * r_val
+    }
+
     return(output)
   }
 
 
-  integrand_m_g_r <- function(a, z, row_data, covars, w_names, q_model, g_model, r_model, exposure, mediator, delta, upper_a) {
+  integrand_m_g_r <- function(a, z, row_data, covars, w_names, q_model, g_model, r_model, exposure, mediator, delta, upper_a, density_type) {
     rep <- dim(z)[2]
     row_data <- do.call("rbind", replicate(rep, row_data, simplify = FALSE))
     new_data_m <- new_data_g <- new_data_r <- row_data
@@ -88,6 +94,58 @@ integrate_psi_g <- function(data, covars, w_names, q_model, r_model, g_model, ex
         outcome = "y"
       )
 
+      m_val <- q_model$predict(task_m)
+
+      if (density_type == "sl") {
+        task_g <- sl3::sl3_Task$new(
+          data = new_data_g,
+          covariates = c(w_names),
+          outcome = exposure,
+        )
+
+        task_r <- sl3::sl3_Task$new(
+          data = new_data_r,
+          covariates = c(w_names),
+          outcome = mediator,
+        )
+
+        g_val <- g_model$predict(task_g)
+        r_val <- r_model$predict(task_r)
+
+        output <- m_val * g_val$likelihood * r_val$likelihood
+      } else {
[…]
