+++ sandbox/04_make_plots.R
@@ -6,31 +6,33 @@ library(cividis)
 library(hrbrthemes)
 
 sim_results_1 <- readRDS(
-  here("sandbox/data/SuperNOVA_mediation_cont_mc.rds")
+  here("sandbox/data/SuperNOVA_mediation_quant_1.rds")
 )
 
-# sim_results_2 <- readRDS(
-#   here("sandbox/data/SuperNOVA_2_sim.rds")
-# )
+sim_results_2 <- readRDS(
+  here("sandbox/data/SuperNOVA_mediation_quant_2.rds")
+)
+
+sim_results_3 <- readRDS(
+  here("sandbox/data/SuperNOVA_mediation_quant_3.rds")
+)
 #
-# sim_results_3 <- readRDS(
-#   here("sandbox/data/SuperNOVA_3_sim.rds")
-# )
-#
-# sim_results_4 <- readRDS(
-#   here("sandbox/data/SuperNOVA_5_sim.rds")
-# )
-#
-# sim_results_5 <- readRDS(
-#   here("sandbox/data/SuperNOVA_6_sim.rds")
-# )
+sim_results_4 <- readRDS(
+  here("sandbox/data/SuperNOVA_mediation_quant_4.rds")
+)
+
+sim_results_5 <- readRDS(
+  here("sandbox/data/SuperNOVA_mediation_quant_5.rds")
+)
+
+
 #
 # sim_results_6 <- readRDS(
 #   here("sandbox/data/SuperNOVA_7_sim.rds")
 # )
 
 
-sim_results <- rbind(sim_results_1)
+sim_results <- rbind(sim_results_1, sim_results_2, sim_results_3, sim_results_4, sim_results_5)
 
 
 sim_statistics <- sim_results %>%
@@ -54,7 +56,12 @@ sim_statistics <- sim_results %>%
     mean_nie_bias_int = mean(integrated_nie_bias),
     nie_est_sd_int = sd(integrated_nie_est),
     nie_est_MSE_int = mean_nie_bias_int^2 + nie_est_sd_int^2,
-    nie_int_CI_coverage = mean(integrated_nie_cov)
+    nie_int_CI_coverage = mean(integrated_nie_cov),
+    #
+    mean_ate_bias = mean(ate_bias),
+    ate_est_sd = sd(ate_est),
+    ate_est_MSE = mean_ate_bias^2 + ate_est_sd^2,
+    ate_CI_coverage = mean(ate_cov)
     #
   ) %>%
   mutate(
@@ -62,16 +69,20 @@ sim_statistics <- sim_results %>%
     abs_nde_int_bias = abs(mean_nde_bias_int),
     abs_nie_pseudo_bias = abs(mean_nie_bias_pseudo),
     abs_nie_int_bias = abs(mean_nie_bias_int),
+    abs_ate_bias = abs(mean_ate_bias),
     #
     nde_pseudo_sqrt_n_abs_bias = sqrt(n_obs) * abs_nde_pseudo_bias,
     nde_int_sqrt_n_abs_bias = sqrt(n_obs) * abs_nde_int_bias,
     nie_pseudo_sqrt_n_abs_bias = sqrt(n_obs) * abs_nie_pseudo_bias,
     nie_int_sqrt_n_abs_bias = sqrt(n_obs) * abs_nie_int_bias,
+    ate_sqrt_n_abs_bias = sqrt(n_obs) * abs_ate_bias,
     #
     nde_pseudo_n_MSE = n_obs * nde_est_MSE_pseudo,
     nde_int_n_MSE = n_obs * nde_est_MSE_int,
     nie_pseudo_n_MSE = n_obs * nie_est_MSE_pseudo,
[…]
