+++ sandbox/03_run_med_simulation_cont_mc_9.R
@@ -87,5 +87,5 @@ for (sample_size in n_obs) {
 # save results to file
 saveRDS(
   object = sim_results_df,
-  file = here("sandbox/data", paste0("SuperNOVA_", "mediation_cont_mc_1", ".rds"))
+  file = here("sandbox/data", paste0("SuperNOVA_", "mediation_cont_mc_9", ".rds"))
 )
