+++ R/indiv_stoch_shift_est_g_exp.R
@@ -262,158 +262,126 @@ indiv_stoch_shift_est_g_exp <- function(exposure,
       if (max_hn > hn_trunc_thresh) {
         delta_reduced <- delta_reduced - (0.1 * delta_reduced)
         # need a data set with the exposure stochastically shifted DOWNWARDS A-delta
-        av_downshifted <- data.table::copy(av)
 
-        data.table::set(av_downshifted,
-          j = exposure,
-          value = shift_additive(
-            a = subset(av, select = exposure),
-            delta = -delta_reduced,
-            lower_bound = lower_bound,
-            upper_bound = upper_bound
+        av_downshifted <- create_shifted_data(av, exposure, -delta_reduced, lower_bound, upper_bound)
+        at_downshifted <- create_shifted_data(at, exposure, -delta_reduced, lower_bound, upper_bound)
+        av_upshifted <- create_shifted_data(av, exposure, delta_reduced, lower_bound, upper_bound)
+        at_upshifted <- create_shifted_data(at, exposure, delta_reduced, lower_bound, upper_bound)
+        av_upupshifted <- create_shifted_data(av, exposure, 2 * delta_reduced, lower_bound, upper_bound)
+        at_upupshifted <- create_shifted_data(at, exposure, 2 * delta_reduced, lower_bound, upper_bound)
+
+        if (density_type == "sl") {
+
+          sl_task <- sl3::sl3_Task$new(
+            data = at,
+            outcome = exposure,
+            covariates = covars
           )
-        )
 
-        # need a data set with the exposure stochastically shifted DOWNWARDS A-delta
-        av_downshifted <- data.table::copy(av)
-
-        data.table::set(av_downshifted,
-          j = exposure,
-          value = shift_additive(
-            a = subset(av, select = exposure),
-            delta = -delta_reduced,
-            lower_bound = lower_bound,
-            upper_bound = upper_bound
+          sl_task_noshift_at <- sl3::sl3_Task$new(
+            data = at,
+            outcome = exposure,
+            covariates = covars
           )
-        )
 
-        at_downshifted <- data.table::copy(at)
-
-        data.table::set(at_downshifted,
-          j = exposure,
-          value = shift_additive(
-            a = subset(at, select = exposure),
-            delta = -delta_reduced,
-            lower_bound = lower_bound,
-            upper_bound = upper_bound
+          sl_task_noshift_av <- sl3::sl3_Task$new(
+            data = av,
+            outcome = exposure,
+            covariates = covars
           )
-        )
 
-        # need a data set with the exposure stochastically shifted UPWARDS A+delta
-        av_upshifted <- data.table::copy(av)
-        data.table::set(av_upshifted,
-          j = exposure,
-          value = shift_additive(
-            a = subset(av, select = exposure),
-            delta = delta_reduced,
-            lower_bound = lower_bound,
-            upper_bound = upper_bound
+          # sl3 task for data with exposure shifted DOWNWARDS A-delta
+          sl_task_downshifted_at <- sl3::sl3_Task$new(
+            data = at_downshifted,
+            outcome = exposure,
+            covariates = covars
           )
-        )
 
-        at_upshifted <- data.table::copy(at)
-        data.table::set(at_upshifted,
-          j = exposure,
-          value = shift_additive(
-            a = subset(at, select = exposure),
-            delta = delta_reduced,
-            lower_bound = lower_bound,
[…]
