+++ R/SuperNOVA.R
@@ -461,9 +461,9 @@ SuperNOVA <- function(w,
             exposures,
             deltas,
             g_learner = pi_learner,
-            covars,
-            av,
-            at,
+            covars = covars,
+            av = av,
+            at = at,
             adaptive_delta = adaptive_delta,
             hn_trunc_thresh = hn_trunc_thresh,
             upper_bound = upper_bound,
