+++ R/final_result_utils.R
@@ -36,8 +36,9 @@ calc_final_ind_shift_param <- function(tmle_fit, exposure, fold_k) {
 
 #' Calculates the Effect Modification Shift Parameter
 #'
-#' @description This function uses a decision tree estimator to regress the efficient influence function (EIF)
+#' @description This function uses a decision tree estimator to regress the difference in expected outcomes
 #'  of an exposure onto a covariate. Both the exposure and the covariate have been determined through data-adaptive methods.
+#'  If no rules are found, a median split is applied.
 #'
 #' @param tmle_fit_av TMLE results for the validation fold.
 #' @param tmle_fit_at TMLE results for the training fold.
@@ -61,154 +62,141 @@ calc_final_effect_mod_param <- function(tmle_fit_av,
                                         effect_m_name,
                                         fold_k,
                                         em_learner) {
+  # Calculate pseudo_outcome for training and validation datasets
   pseudo_outcome_at <- tmle_fit_at$qn_shift_star - tmle_fit_at$qn_noshift_star
   pseudo_outcome_av <- tmle_fit_av$qn_shift_star - tmle_fit_av$qn_noshift_star
 
+  # Rename columns
   names(pseudo_outcome_at) <- "pseudo_outcome_at"
   names(pseudo_outcome_av) <- "pseudo_outcome_av"
 
+  # Prepare data for the tree model
   em_model_data_at <- cbind.data.frame(
     pseudo_outcome_at, subset(at,
-      select = effect_m_name
+                              select = effect_m_name
     )
   )
 
   em_model_data_av <- cbind.data.frame(
     pseudo_outcome_av, subset(av,
-      select = effect_m_name
+                              select = effect_m_name
     )
   )
 
-  task <- sl3::make_sl3_Task(
-    data = em_model_data_at,
-    covariates = effect_m_name,
-    outcome = "pseudo_outcome_at",
-    outcome_type = "continuous"
-  )
+  # Create the formula for the tree model
+  response_var <-  "pseudo_outcome_at"
+  formula_string <- paste(response_var, "~", effect_m_name)
+  formula <- as.formula(formula_string)
 
-  discrete_sl_metalrn <- sl3::Lrnr_cv_selector$new(sl3::loss_squared_error)
+  # Fit the tree model
+  tree_model <- glmtree(formula, data = em_model_data_at)
 
-  discrete_sl <- sl3::Lrnr_sl$new(
-    learners = em_learner,
-    metalearner = discrete_sl_metalrn,
-  )
+  # Extract the rules from the tree model
+  rules <- list_rules_party(tree_model)
 
-  sl_fit <- suppressWarnings(discrete_sl$train(task))
-
-  selected_learner <- sl_fit$learner_fits[[
-    which(sl_fit$coefficients == 1)
-  ]]
-
-
-  rules <- list_rules_party(selected_learner$fit_object)
-
+  # If no rules are found, use median split
   if (all(rules == "") == TRUE) {
-    # create a rule for the median -----
-
-    if (all(apply(at[, ..effect_m_name], 2, function(x) {
+    # Determine if the effect modifier is binary (0 or 1)
+    is_binary <- all(apply(at[, ..effect_m_name], 2, function(x) {
       all(x %in% 0:1)
-    })) == TRUE) {
-      medians <- apply(at[, ..effect_m_name], 2, median)
+    }))
 
-      median_rule_list <- list()
-      for (i in seq(length(medians))) {
-        median_rule <- paste(names(medians)[i], "==", medians[i])
[…]
