+++ sandbox/slurm_scripts/03_run_med_simulation_cont_quad_1.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 # Job name:
-#SBATCH --job-name=c_quad_mediation_supernova
+#SBATCH --job-name=1_c_quad_mediation_supernova
 #
 # Partition:
 #SBATCH --partition=savio3
@@ -22,4 +22,4 @@
 module load r/4.0.3
 
 ### Run Simulation
-R CMD BATCH --no-save ../03_run_med_simulation_cont_quad.R supernova_mediation_cont_quad.Rout
+R CMD BATCH --no-save ../03_run_med_simulation_cont_quad_1.R supernova_mediation_cont_quad_1.Rout
