### 2023-04-25 ec79fd6
updates to not just have multinomial for discrete data but also density for each bin.
