+++ R/integrate_psi_g_discrete.R
@@ -27,7 +27,7 @@
 #'  outcome mechanism at the natural value of the exposure Q(A, W) and an
 #'  upshift of the exposure Q(A + delta, W)
 
-integrate_psi_g_discrete <- function(av, at, covars, w_names, q_model, r_model, g_model, exposure, mediator, delta, n_bins, n_samples = 1000, method = "MC", mediator_quantized) {
+integrate_psi_g_discrete <- function(av, at, covars, w_names, q_model, r_model, g_model, exposure, mediator, delta, n_bins, n_samples = 1000, method = "MC", mediator_quantized, density_type, upper_bound, use_multinomial) {
   av <- as.data.frame(av)
   at <- as.data.frame(at)
 
@@ -36,12 +36,12 @@ integrate_psi_g_discrete <- function(av, at, covars, w_names, q_model, r_model, 
 
   sample_z <- runif(n_samples, min = lower_z, max = upper_z)
 
-  integrand_m_r <- function(sample_z_inner, row_data, covars, w_names, q_model, r_model, exposure, mediator, delta, upper_a, scale_data, mediator_quantized) {
+  integrand_m_r <- function(sample_z_inner, row_data, covars, w_names, q_model, r_model, exposure, mediator, delta, upper_bound, density_type, use_multinomial) {
     row_data_rep <- do.call("rbind", replicate(length(sample_z_inner), row_data, simplify = FALSE))
     new_data_m <- new_data_r <- row_data_rep
 
     new_data_m[mediator] <- sample_z_inner
-    new_data_m[exposure] <- ifelse(new_data_m[[exposure]] + delta >= upper_a, upper_a, new_data_m[[exposure]] + delta)
+    new_data_m[exposure] <- ifelse(new_data_m[[exposure]] + delta >= upper_bound, upper_bound, new_data_m[[exposure]] + delta)
 
     new_data_r[mediator] <- sample_z_inner
 
@@ -51,32 +51,38 @@ integrate_psi_g_discrete <- function(av, at, covars, w_names, q_model, r_model, 
       outcome = "y"
     )
 
-    task_r <- sl3::sl3_Task$new(
-      data = new_data_r,
-      covariates = c(w_names),
-      outcome = mediator,
-    )
-
     m_val <- q_model$predict(task_m)
 
-    if (mediator_quantized) {
-      r_val <- r_model$predict(task_r)
-      r_vals <- unlist(r_val, recursive = FALSE)
-      likelihood_r <- sapply(r_vals, function(x) x[[sample_z_inner]])
-      output <- m_val * likelihood_r
-    } else {
-      r_val <- r_model$predict(task_r)$likelihood
+    if (density_type == "sl") {
+      task_r <- sl3::sl3_Task$new(
+        data = new_data_r,
+        covariates = c(w_names),
+        outcome = mediator,
+      )
+
+      if (use_multinomial) {
+        r_val <- r_model$predict(task_r)
+        r_vals <- unlist(r_val, recursive = FALSE)
+        likelihood_r <- sapply(r_vals, function(x) x[[sample_z_inner]])
+        output <- m_val * likelihood_r
+      } else {
+        r_val <- r_model$predict(task_r)$likelihood
+        output <- m_val * r_val
+      }
+    }else{
+      r_val <- suppressMessages(predict(r_model, new_A = new_data_r[[mediator]], new_W = new_data_r[w_names]))
       output <- m_val * r_val
     }
+
     return(output)
   }
 
-  integrand_m_g_r_mc <- function(sample_a, sample_z, row_data, covars, w_names, q_model, g_model, r_model, exposure, mediator, delta, upper_a, scale_data, mediator_quantized) {
[…]
