### 2023-04-28 #5170 @csoneson
👋🏻 @lucasmiranda42, @yafeiwang89 - would one of you be interested in reviewing this submission for JOSS?

> SuperNOVA: Semi-Parametric Identification and Estimation of Interaction and Effect Modification in Mixed Exposures using Stochastic Interventions in R
