### 2023-04-29 blind-contours/SuperNOVA#1 @lucasmiranda42
Installation instructions
Hey there! Lucas from [5422](https://github.com/openjournals/joss-reviews/issues/5422) here.

I just installed the package and everything ran smoothly. Looking forward to trying it out in the next few days!

While the installation instructions proceed as expected, I wonder if it would make sense to warn people they should install remotes too. It should be fairly obvious to most people installing a custom R package, but it's nevertheless a requirement that can be easily overlooked when creating a new environment, for example.

Cheers!
Lucas
