### 2023-05-01 blind-contours/SuperNOVA#2 @lucasmiranda42
Problem with simulate_data when trying to reproduce the provided example
Hello again!  Another issue from [5422](https://github.com/openjournals/joss-reviews/issues/5422) here.

I run into a problem when trying to run the provided example. Upon excecuting

`simulate_data(n_obs = n_obs, shift_var_index = c(3))`

I get an error:

`Error in simulate_data(n_obs = n_obs, shift_var_index = c(3)) :
  unused argument (shift_var_index = c(3))`

Since the parameter `shift_var_index` does not seem to be part of the API anymore. Please clarify or update the documentation.

Best!
Lucas
