### 2023-05-01 #5422 @lucasmiranda42
Hi! First impressions on my side.

1) The paper is well written: it clearly states how the package fits in the field, and the problems it intends to solve. My only concern is regarding the references: in the README file of the main repo there are several other references that are not mentioned in the manuscript. I would suggest either adding them or justifying why this is not the case :)

2) The package installs without problems. I filed an issue with a minor concern regarding how the instructions are documented, but overall everything seems to work properly. The example provided in the README file does however not work. It seems to be due to an undocumented change in the API; I filed an issue accordingly.

3) While automatic tests are provided, the CI pipeline seems to be failing at the moment. I would suggest inspecting why and fixing the failing tests.

4) Lastly, while the provided example seems well documented (I haven't gone fully through it yet due to the issue mentioned above) I couldn't find a link to more extensive documentation of the package capabilities and functions. In case it indeed does not exist, I would suggest expanding here!

I will for now wait until (2) is clarified and/or fixed, and fully test functionality then.

Best, and looking forward!
Lucas
