### 2023-05-05 2dba8df
Check passing except weird issue with multicore in vignette- rewrite of readme.
