+++ R/integrate_psi_aw_g_quant.R
@@ -2,16 +2,18 @@
 #'
 #' @details Does the double integration as described in lemma 1
 #'
-#' @param data A \code{character} vector of exposures to be shifted.
-#' @param covars The mediator variable
-#' @param w_names Covariate names
-#' @param q_model A \code{character} vector covariates to adjust for.
-#' @param g_model The training data
+#' @param at Training data
+#' @param av Validation data
+#' @param covars Covariates using the m model
+#' @param w_names Covariates used in the g model
+#' @param pseudo_model Pseudo regression model
+#' @param g_model Exposure density/probability model
 #' @param exposure A \code{numeric} indicating the magnitude of the shift to be
 #'  computed for the exposure \code{A}. This is passed to the internal
-#' @param delta Object containing a set of instantiated learners from the
-#'  \pkg{sl3}, to be used in fitting an ensemble model.
-#'
+#' @param psi_aw Vector of pseudo regression predictions
+#' @param n_bins Number of bins in the discrete exposure
+#' @param use_multinomial TRUE/FALSE whether multinomial regression is used
+#' @param density_type Type of density estimation sl or hal
 #' @importFrom stats glm as.formula predict
 #' @importFrom data.table as.data.table setnames copy set
 #' @importFrom stringr str_detect
@@ -20,11 +22,11 @@
 #' @return A \code{data.table} with two columns, containing estimates of the
 #'  outcome mechanism at the natural value of the exposure Q(A, W) and an
 #'  upshift of the exposure Q(A + delta, W).
-integrate_psi_aw_g_quant <- function(at, av, covars, w_names, pseudo_model, g_model, exposure, psi_aw, n_bins, use_multinomial) {
+integrate_psi_aw_g_quant <- function(at, av, covars, w_names, pseudo_model, g_model, exposure, psi_aw, use_multinomial, density_type) {
   at <- as.data.frame(at)
   av <- as.data.frame(av)
 
-  integrand_psi_g_quant <- function(bin_a, row_data, covars, pseudo_model, g_model, exposure, av, use_multinomial) {
+  integrand_psi_g_quant <- function(bin_a, row_data, covars, w_names, pseudo_model, g_model, exposure, av, use_multinomial, density_type) {
     row_data <- do.call("rbind", replicate(length(bin_a), row_data, simplify = FALSE))
     new_data_m <- new_data_g <- row_data
     new_data_m[exposure] <- bin_a
@@ -48,7 +50,7 @@ integrate_psi_aw_g_quant <- function(at, av, covars, w_names, pseudo_model, g_mo
 
       if (use_multinomial) {
         g_val <- g_model$predict(task_g)
-        index <- ifelse(bin_a + delta >= upper, upper, bin_a + delta)
+        index <- bin_a
         g_val <- unlist(g_val)[[index]]
       }else{
         g_val <- g_model$predict(task_g)$likelihood
@@ -67,7 +69,7 @@ integrate_psi_aw_g_quant <- function(at, av, covars, w_names, pseudo_model, g_mo
 
   for (i in 1:nrow(av)) {
     row_data <- av[i, ]
-    integral_values <- integrand_psi_g_quant(bin_a = unique(av[[exposure]]), row_data, covars, pseudo_model, g_model, exposure, av, use_multinomial)
+    integral_values <- integrand_psi_g_quant(bin_a = unique(av[[exposure]]), row_data, covars, w_names, pseudo_model, g_model, exposure, av, use_multinomial, density_type)
 
     integral_result <- sum(integral_values)
 
