+++ tests/testthat/test-default_sl_construction.R
@@ -19,4 +19,3 @@ expect_true(class(sls$zeta_learner)[1] == "Stack")
 expect_true(class(sls$mu_learner)[1] == "Stack")
 expect_true(class(sls$e_learner)[1] == "Stack")
 expect_true(class(sls$g_learner)[1] == "Stack")
-expect_true(class(sls$em_learner)[1] == "Stack")

[…]
