### 2023-05-05 251b023
Prepare slurm files for mediation simulation.
