### 2023-05-05 3c606be
update the functions with use_multinomial parameter.
