+++ sandbox/03_run_med_simulation_quant_3.R
@@ -55,7 +55,7 @@ for (sample_size in n_obs) {
     z <- data_sim[, mediators]
     y <- data_sim[, outcome]
 
-    est_out_discrete_m <- fit_estimators_mediation(
+    est_out_discrete_e <- fit_estimators_mediation(
       w = w,
       a = a,
       z = z,
@@ -70,14 +70,16 @@ for (sample_size in n_obs) {
       var_sets = "a-z",
       exposure_quantized = TRUE,
       mediator_quantized = FALSE,
-      n_mc_sample = n_mc_sample,
+      n_mc_sample = 2*sample_size,
       density_type = "sl",
-      integration_method = "MC"
+      integration_method = "MC",
+      n_bins = 10,
+      use_multinomial = FALSE
     )
 
-    est_out_discrete_m$n_obs <- sample_size
+    est_out_discrete_e$n_obs <- sample_size
 
-    results[[this_iter]] <- est_out_discrete_m
+    results[[this_iter]] <- est_out_discrete_e
   }
   # concatenate iterations
   results_out <- bind_rows(results, .id = "sim_iter")
