### 2023-05-08 blind-contours/SuperNOVA#2 @blind-contours
Hi @lucasmiranda42 - thanks for your review of the package. I've since updated this function and improved documentation in the README. Now I just build in the shift index and in the README I show correct estimation for each parameter.
