### 2023-05-08 blind-contours/SuperNOVA#1 @blind-contours
I'll add a little section in the README to discuss remotes.
