### 2023-05-08 blind-contours/SuperNOVA#2 @lucasmiranda42
Wonderful! I'll close and reopen if needed
