### 2023-05-08 blind-contours/SuperNOVA#1 @lucasmiranda42
Thank you! I'll close the issue then
