### 2023-05-08 blind-contours/SuperNOVA#3 @lucasmiranda42
Issue running the provided example
Hello again! Lucas from [5422](https://github.com/openjournals/joss-reviews/issues/5422)

I just tried the provided example and ran into another issue, likely of the same nature as the last one:

When running:

`sim_results <- SuperNOVA(
    w = w,
    a = a,
    y = y,
    delta = deltas,
    n_folds = 3,
    num_cores = 6,
    outcome_type = "continuous",
    quantile_thresh = 0,
    seed = 294580
)`

I get this error:

`Error in (function (.x, .f, ..., .progress = FALSE)  :
  i In index: 1.
Caused by error in indiv_stoch_shift_est_g_exp():
! unused argument (exposure_quantized = exposure_quantized)`

Am I doing anything wrong? Or is it still an error on the provided code?

Best, and thank you for addressing the previous comments,
Lucas
