### 2023-05-09 #5422 @mhu48
Hi,

I have reviewed your paper titled "Package X for Statistical Analysis" submitted to the Journal of Open Statistical Software. Overall, the paper is well written and clearly articulates the purpose and scope of the proposed package. However, I do have some concerns and suggestions that I would like to share with you.

Firstly, I recommend that you either include some more references in the manuscript to support this research project, or justify why the current two listed are sufficient for the paper.

Secondly, I successfully installed the package without any issues. However, I encountered similar problems with the example provided in the README file and the CI pipeline. It would be a good idea if you could investigate and address the issues causing the failures to ensure that the package is reliable and robust. I look forward to reviewing the full functionality and performance of the package once this is resolved. I also suggest adding one or two more alternative examples so the new user could more easily understand the different ways to apply this package.

Lastly, I would appreciate more details discussing the potential scope and usage of the package in the second paragraph under the secion, "SuperNOVA ’s Scope".

Overall, I believe that the package has the potential to be a valuable contribution to the field.

Best regards,
