+++ paper/paper.bib
@@ -1,54 +1,3 @@
-@article{Bobb2014,
-abstract = {Because humans are invariably exposed to complex chemical mixtures, estimating the health effects of multi-pollutant exposures is of critical concern in environmental epidemiology, and to regulatory agencies such as the U.S. Environmental Protection Agency. However, most health effects studies focus on single agents or consider simple two-way interaction models, in part because we lack the statistical methodology to more realistically capture the complexity of mixed exposures. We introduce Bayesian kernel machine regression (BKMR) as a new approach to study mixtures, in which the health outcome is regressed on a flexible function of the mixture (e.g. air pollution or toxic waste) components that is specified using a kernel function. In high-dimensional settings, a novel hierarchical variable selection approach is incorporated to identify important mixture components and account for the correlated structure of the mixture. Simulation studies demonstrate the success of BKMR in estimating the exposure-response function and in identifying the individual components of the mixture responsible for health effects. We demonstrate the features of the method through epidemiology and toxicology applications.},
-author = {Bobb, Jennifer F. and Valeri, Linda and {Claus Henn}, Birgit and Christiani, David C. and Wright, Robert O. and Mazumdar, Maitreyi and Godleski, John J. and Coull, Brent A.},
-doi = {10.1093/biostatistics/kxu058},
-file = {:Users/davidmccoy/Library/Application Support/Mendeley Desktop/Downloaded/Bobb et al. - 2014 - Bayesian kernel machine regression for estimating the health effects of multi-pollutant mixtures.pdf:pdf},
-issn = {14684357},
-journal = {Biostatistics},
-keywords = {Air pollution,Bayesian variable selection,Environmental health,Gaussian process regression,Metal mixtures.},
-number = {3},
-pages = {493--508},
-pmid = {25532525},
-title = {{Bayesian kernel machine regression for estimating the health effects of multi-pollutant mixtures}},
-volume = {16},
-year = {2014}
-}
-@article{Keil2019,
[…]
