+++ R/integrand_q_g.R
@@ -17,8 +17,8 @@
 integrand_q_g <- function(sample_a, row_data, covars, w_names, q_model, g_model, exposure, g_delta, m_delta, upper_bound, density_type) {
   row_data <- do.call("rbind", replicate(length(sample_a), row_data, simplify = FALSE))
   new_data_m <- new_data_g <- row_data
-  new_data_m[exposure] <- ifelse(sample_a + m_delta >= upper_bound, upper_bound, sample_a + m_delta)
-  new_data_g[exposure] <- ifelse(sample_a + g_delta >= upper_bound, upper_bound, sample_a + g_delta)
+  new_data_m[[exposure]] <- ifelse(sample_a + m_delta >= upper_bound, upper_bound, sample_a + m_delta)
+  new_data_g[[exposure]] <- ifelse(sample_a + g_delta >= upper_bound, upper_bound, sample_a + g_delta)
 
   task_m <- sl3::sl3_Task$new(
     data = new_data_m,
