+++ sandbox/Project_Viva/ForDavid.Rmd
@@ -1,332 +1,0 @@
----
-title: "Prenatal metals (1st trimester) and mitochondrial DNA copy number and telomere length in maternal second trimester and cord blood"
-author: "Anna Smith"
-date: "09/27/2021"
-output: html_document
-
-# Created: 09/27/2021 
-# Updated: 
-# Includes: Data for CVtreemle
----
-
-```{r globals}
-library(SuperNOVA)
-library(here)
-library(readr)
-library(tidyverse)
-seed <- 11122441
-```
-#prenatal metal mixture and cord blood mitochondrial DNA copy number (mtDNAcn)
-
-```{r cord blood mtDNA}
-
-mtDNA_birth<- read_csv(here("sandbox/Project_Viva/mtDNA_birth.csv"))
-
-#Recode plate number to factor variable
-mtDNA_birth$dna_copyn_assay_plate_child<- as.factor(
-  mtDNA_birth$dna_copyn_assay_plate_child)
-
-table(mtDNA_birth$dna_copyn_assay_plate_child)
-
-#Recode race to factor variable
-mtDNA_birth$race2_mom_epi_epia_d<- as.factor(
-  mtDNA_birth$race2_mom_epi_epia_d)
-
-table(mtDNA_birth$race2_mom_epi_epia_d)
-
-#recode other variables to factor variables
-mtDNA_birth$coll_grad<- as.factor(mtDNA_birth$coll_grad)
-mtDNA_birth$gt70k<- as.factor(mtDNA_birth$gt70k)
-mtDNA_birth$smokpreg_final_d<- as.factor(mtDNA_birth$smokpreg_final_d)
-mtDNA_birth$nullip<- as.factor(mtDNA_birth$nullip)
-mtDNA_birth$female_d<- as.factor(mtDNA_birth$female_d)
-
-#save the names of the mixture variables in the variable "Xnm"
-#note: non log2-transformed metals = "As" "Ba", etc. if want to use those instead
-exposure = c("As", "Ba", "Cd", "Cs", "Hg", "Mg", 
-             "Mn", "Pb", "Se", "Zn")
-
-deltas = c("As" = 1, "Ba" = 1, "Cd" = 1, "Cs" = 1 , "Hg" = 1, "Mg" = 1, 
-             "Mn" = 1, "Pb" = 1, "Se" = 1, "Zn" = 1)
-
-exposures <- mtDNA_birth[exposure]
-
-mediator <- mtDNA_birth["dna_copyn_mean_child"]
-
-#log2 transform the outcome
-#note: non log2-transformed outcomes = original value 
-mtDNA_birth$dna_copyn_mean_child<- log2(mtDNA_birth$dna_copyn_mean_child)
-outcome = mtDNA_birth[["birthweight_d"]]
-
-#save covariates in variable covars
-covars = mtDNA_birth[c('dna_copyn_assay_plate_child', 'age_mom_enroll_d', 
-           'bmi_mom_prepreg_d', 'race2_mom_epi_epia_d', 'coll_grad', 'gt70k', 
-           'smokpreg_final_d', 'nullip', 'female_d')]
-
-mtDNA_birth <- as.data.frame(unclass(mtDNA_birth),                
-                       stringsAsFactors = TRUE)
-
-mtDNA_birth <- mtDNA_birth %>% 
-  mutate_if(is.numeric, ~tidyr::replace_na(.,mean(., na.rm = TRUE)))
-
-
-
-mtDNA_birth_results <- SuperNOVA(w = covars,
-                                 a = exposures,
-                                 z = mediator, 
-                                 y = outcome,
-                                 seed = seed,
-                                 n_folds = 10,
-                                 outcome_type =  "continuous", 
-                                 num_cores = 7,
[…]
