### 2023-05-23 e017258
new shell scripts for adaptive quad
