+++ sandbox/slurm_scripts/03_run_med_simulation_cont_quad_2.sh
@@ -1,0 +1,25 @@
+#!/bin/bash
+# Job name:
+#SBATCH --job-name=2_c_quad_mediation_supernova
+#
+# Partition:
+#SBATCH --partition=savio3
+#
+
+#SBATCH --qos=biostat_savio3_normal
+#SBATCH --account=co_biostat
+
+# Number of nodes for use case:
+#SBATCH --nodes=1
+#SBATCH --ntasks=1
+#SBATCH --cpus-per-task=20
+#SBATCH --exclusive
+#
+# Wall clock limit:
+#SBATCH --time 72:00:00
+#
+## Command(s) to run (example):
+module load r/4.0.3
+
+### Run Simulation
+R CMD BATCH --no-save ../03_run_med_simulation_cont_quad_2.R supernova_mediation_cont_quad_2.Rout
