+++ sandbox/slurm_scripts/03_run_med_simulation_cont_mc_2.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 # Job name:
-#SBATCH --job-name=mc_cont_mediation_supernova_2
+#SBATCH --job-name=2_mc_cont_mediation_supernova
 #
 # Partition:
 #SBATCH --partition=savio3
