+++ sandbox/slurm_scripts/03_run_med_simulation_cont_mc_9.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 # Job name:
-#SBATCH --job-name=mc_cont_mediation_supernova_9
+#SBATCH --job-name=9_mc_cont_mediation_supernova
 #
 # Partition:
 #SBATCH --partition=savio3
