### 2023-05-23 bcfc0fa
fixing sim input
