+++ sandbox/03_run_med_simulation_cont_mc_8.R
@@ -18,7 +18,8 @@ n_fold <- 10
 # Generate simulated data -----------------
 
 full_data <- simulate_mediation_data(n_obs = 100000,
-                                     delta = 1)
+                                     delta = 1,
+                                     n_bins = 10)
 p0_data <- full_data$data
 
 nde_a1 <- full_data$nde_a1
