+++ sandbox/04_make_plots.R
@@ -5,6 +5,48 @@ library(viridis)
 library(cividis)
 library(hrbrthemes)
 
+## Sim results for continuous A using MC integration
+sim_results_1 <- readRDS(
+  here("sandbox/data/SuperNOVA_mediation_cont_mc_1.rds")
+)
+
+sim_results_2 <- readRDS(
+  here("sandbox/data/SuperNOVA_mediation_cont_mc_2.rds")
+)
+
+sim_results_3 <- readRDS(
+  here("sandbox/data/SuperNOVA_mediation_cont_mc_3.rds")
+)
+#
+sim_results_4 <- readRDS(
+  here("sandbox/data/SuperNOVA_mediation_cont_mc_4.rds")
+)
+
+sim_results_5 <- readRDS(
+  here("sandbox/data/SuperNOVA_mediation_cont_mc_5.rds")
+)
+
+sim_results_6 <- readRDS(
+  here("sandbox/data/SuperNOVA_mediation_cont_mc_6.rds")
+)
+
+sim_results_7 <- readRDS(
+  here("sandbox/data/SuperNOVA_mediation_cont_mc_7.rds")
+)
+
+sim_results_8 <- readRDS(
+  here("sandbox/data/SuperNOVA_mediation_cont_mc_8.rds")
+)
+
+sim_results_9 <- readRDS(
+  here("sandbox/data/SuperNOVA_mediation_cont_mc_9.rds")
+)
+
+sim_results_10 <- readRDS(
+  here("sandbox/data/SuperNOVA_mediation_cont_mc_10.rds")
+)
+
+## Sim results for discrete A
 sim_results_1 <- readRDS(
   here("sandbox/data/SuperNOVA_mediation_quant_1.rds")
 )
@@ -26,7 +68,11 @@ sim_results_5 <- readRDS(
 )
 
 
-sim_results <- rbind(sim_results_1, sim_results_2, sim_results_3, sim_results_4)
+sim_results <- rbind(sim_results_1, sim_results_2,
+                     sim_results_3, sim_results_4,
+                     sim_results_5, sim_results_6,
+                     sim_results_7, sim_results_8,
+                     sim_results_9, sim_results_10)
 
 
 sim_statistics <- sim_results %>%
@@ -208,11 +254,11 @@ int_regression_sqrt_n_bias <- make_sim_statistics_plot(sim_statistics_long, stat
 
 # Pseudo-Regression
 pseudo_cov_plot <- make_sim_statistics_plot(sim_statistics_long, names(pseudo_coverage_stats),
-                                     "MSE for In(Direct) Effects", pseudo_coverage_stats)
+                                            title = "MSE for In(Direct) Effects", legend_labels = pseudo_coverage_stats)
 
 # Integration
 integration_cov_plot <- make_sim_statistics_plot(sim_statistics_long, names(int_coverage_stats),
-                                            "Coverage for In(Direct) Effects using Integration", int_coverage_stats)
+                                                 title = "Coverage for In(Direct) Effects using Integration", legend_labels = int_coverage_stats)
 
 # Save the plots to files
 # Save the plots to files
