### 2023-06-07 27f14a5
updates for the path detection simulation.
