+++ R/joint_stoch_shift_est_Q.R
@@ -28,15 +28,18 @@ joint_stoch_shift_est_Q <- function(exposures,
                                     mu_learner,
                                     covars,
                                     av,
-                                    at) {
+                                    at,
+                                    outcome_type) {
   future::plan(future::sequential, gc = TRUE)
 
   # scale the outcome for logit transform
-  y_star_av <- scale_to_unit(vals = av$y)
-  y_star_at <- scale_to_unit(vals = at$y)
+  if (outcome_type != "binary") {
+    y_star_av <- scale_to_unit(vals = av$y)
+    y_star_at <- scale_to_unit(vals = at$y)
 
-  av$y <- y_star_av
-  at$y <- y_star_at
+    av$y <- y_star_av
+    at$y <- y_star_at
+  }
 
   results <- list()
 
