### 2023-06-09 9662252
update to run script to test server after earth fail, updated sl3 algorithm to see if that was the problem and cleaned the NHANES run script.
