+++ sandbox/04_make_plots.R
@@ -1,3 +1,4 @@
+# Load the necessary libraries
 library(tidyverse)
 library(here)
 library(ggpubr)
@@ -5,266 +6,124 @@ library(viridis)
 library(cividis)
 library(hrbrthemes)
 
-
-## simulation results under discrete exposure
-quant_sim_results_1 <- readRDS(
-  here("sandbox/data/SuperNOVA_mediation_quant_1.rds")
-)
-
-quant_sim_results_2 <- readRDS(
-  here("sandbox/data/SuperNOVA_mediation_quant_2.rds")
-)
-
-quant_sim_results_3 <- readRDS(
-  here("sandbox/data/SuperNOVA_mediation_quant_3.rds")
-)
-#
-quant_sim_results_4 <- readRDS(
-  here("sandbox/data/SuperNOVA_mediation_quant_4.rds")
-)
-
-quant_sim_results_5 <- readRDS(
-  here("sandbox/data/SuperNOVA_mediation_quant_5.rds")
-)
-
-## simulation results under continuous exposure using MC integration
-
-cont_sim_mc_results_1 <- readRDS(
-  here("sandbox/data/SuperNOVA_mediation_cont_mc_1.rds")
-)
-
-cont_sim_mc_results_2 <- readRDS(
-  here("sandbox/data/SuperNOVA_mediation_cont_mc_2.rds")
-)
-
-cont_sim_mc_results_3 <- readRDS(
-  here("sandbox/data/SuperNOVA_mediation_cont_mc_3.rds")
-)
-#
-cont_sim_mc_results_4 <- readRDS(
-  here("sandbox/data/SuperNOVA_mediation_cont_mc_4.rds")
-)
-
-cont_sim_mc_results_5 <- readRDS(
-  here("sandbox/data/SuperNOVA_mediation_cont_mc_5.rds")
-)
-
-cont_sim_mc_results_6 <- readRDS(
-  here("sandbox/data/SuperNOVA_mediation_cont_mc_6.rds")
-)
-
-cont_sim_mc_results_7 <- readRDS(
-  here("sandbox/data/SuperNOVA_mediation_cont_mc_7.rds")
-)
-
-cont_sim_mc_results_8 <- readRDS(
-  here("sandbox/data/SuperNOVA_mediation_cont_mc_8.rds")
-)
-
-cont_sim_mc_results_9 <- readRDS(
-  here("sandbox/data/SuperNOVA_mediation_cont_mc_9.rds")
-)
-
-cont_sim_mc_results_10 <- readRDS(
-  here("sandbox/data/SuperNOVA_mediation_cont_mc_10.rds")
-)
-
-
-cont_sim_mc_results <- rbind(cont_sim_mc_results_1, cont_sim_mc_results_2, cont_sim_mc_results_3, cont_sim_mc_results_4, cont_sim_mc_results_5, cont_sim_mc_results_6, cont_sim_mc_results_7, cont_sim_mc_results_8, cont_sim_mc_results_9, cont_sim_mc_results_10)
-
-data_to_plot <- cont_sim_mc_results
-
-
-sim_statistics <- data_to_plot %>%
-  group_by(n_obs) %>%
-  summarize(
-    mean_nde_bias_pseudo = mean(pseudo_reg_nde_bias),
-    nde_est_sd_pseudo = sd(pseudo_reg_nde_est),
-    nde_est_MSE_pseudo = mean_nde_bias_pseudo^2 + nde_est_sd_pseudo^2,
-    nde_pseudo_CI_coverage = mean(pseudo_reg_nde_cov),
-    #
-    mean_nde_bias_int = mean(integrated_nde_bias),
[…]
