+++ R/simulate_complicated_mediation_data.R
@@ -2,7 +2,6 @@
 #'
 #' @param n_obs Number of observations
 #' @param delta Delta for M1 Exposure
-#' @param breaks Number of bins to discretize exposure
 #'
 #' @return A data frame of simulated mediation data
 #' @export
