+++ R/utils_create_sls.R
@@ -21,10 +21,10 @@ create_sls <- function() {
   make_continuous_superlearner <- function() {
     learners <- c(
       sl3::Lrnr_glm$new(),
-      sl3::Lrnr_glmnet$new(alpha = 1)
+      sl3::Lrnr_glmnet$new(alpha = 1),
       # sl3::Lrnr_glmnet$new(alpha = 0),
       # sl3::Lrnr_glmnet$new(alpha = .5),
-      # sl3::Lrnr_ranger$new(num.trees = 100),
+      sl3::Lrnr_ranger$new(num.trees = 100)
       # sl3::Lrnr_ranger$new(num.trees = 500),
       # sl3::Lrnr_xgboost$new(nrounds = 50),
       # sl3::Lrnr_xgboost$new(nrounds = 200),
@@ -147,12 +147,12 @@ lrnr_xgboost_300 <- make_learner(Lrnr_xgboost, nrounds = 300)
 
   learners <- c(
     lrnr_glm_basic,
-    lrnr_ridge,
-    lrnr_lasso,
-    lrnr_ranger_100
-    # lrnr_xgboost_50,
+    # lrnr_ridge,
+    # lrnr_lasso,
+    lrnr_ranger_100,
+    lrnr_xgboost_50,
     # lrnr_xgboost_100,
-    # lrnr_xgboost_200,
+    lrnr_xgboost_200
     # lrnr_xgboost_300
   )
 
