### 2023-08-28 #5422 @blind-contours
@csoneson, @mhu48, @lucasmiranda42 -- I've fixed the example in the readme introduction and re-written the readme to be clearer and include mediation. I've also updated the paper. The current CI is still broken because SL3 has an imputation dependency that isn't on CRAN anymore so the install fails at the moment. I'm trying to figure out how to fix this in the YML script. But for now, everything in the package should work. Please let me know what other comments/suggestions you have and I'll update the package.
