### 2023-10-09 #5422 @lucasmiranda42
Dear all,
the package now runs smoothly on my end and fulfils all functionality claims. My only concern is regarding documentation, which I couldn't find (other than the README). @blind-contours could you elaborate on whether this is intended and why API docs may not be needed?
With this, the paper would have green light on my side as well.
Best!
Lucas
