### 2023-10-09 9d0253a
rebuilding vignette and adding docs file
