+++ doc/SuperNOVA-vignette.Rmd
@@ -1,0 +1,241 @@
+---
+title: "Analysis of Variance using Super Learner with Data-Adaptive Stochastic Interventions"
+author: "David McCoy"
+date: "`r Sys.Date()`"
+output: rmarkdown::html_vignette
+bibliography: ../inst/references.bib
+vignette: >
+  %\VignetteIndexEntry{Analysis of Variance using Super Learner with Data-Adaptive Stochastic Interventions}
+  %\VignetteEngine{knitr::rmarkdown}
+  %\VignetteEncoding{UTF-8}
+---
+
+```{r, include = FALSE}
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>"
+)
+```
+
+## Motivation 
+
+The motivation behind the package SuperNOVA is to address the limitations of traditional statistical methods in environmental epidemiology studies. Analysts are often interested in understanding the joint impact of a mixed exposure, i.e. a vector of exposures, but the most important variables and variable sets remain unknown. Traditional methods make overly simplistic assumptions, such as linear and additive relationships, and the resulting statistical quantities may not be directly applicable to public policy decisions.
+
+To overcome these limitations, SuperNOVA uses data-adaptive machine learning methods to identify the variables and variable sets that have the most explanatory power on an outcome of interest. The package builds a discrete Super Learner, which is then analyzed using ANOVA style analysis to determine the variables that contribute most to the model fit through basis functions. The target parameter, which may be a single shift, effect modification, or interaction, is then applied to the data-adaptively determined variable sets.
+
+In this way, SuperNOVA allows analysts to explore modified treatment policies and ask causal questions, such as "If exposure to collections of PFAS chemicals increases, what is the change in cholesterol, immune function, or cancer?" By providing more nuanced and data-driven insights, SuperNOVA aims to inform public policy decisions and improve our understanding of the impact of mixed exposures on health outcomes.
+
+### Data-Adaptive Machine Learning
+
+The package SuperNOVA uses a data-adaptive approach to identify important variables and variable sets in mixed exposure studies. To avoid over-fitting and incorrect model assumptions, the package employs cross-validation procedures. In each fold, the training data is used to:
+
+1. Identify the variable sets of interest using flexible machine learning algorithms.
+2. Fit estimators for the relevant nuisance parameters and the final target parameter of interest using the identified variable sets.
+3. Obtain estimates of the nuisance parameters and the final target parameter of interest using the held-out validation data.
+4. The process is repeated in each fold to provide an estimate specific to each validation data. To optimize the balance between bias and variance, SuperNOVA uses targeted minimum loss based estimation (TMLE).
+5. Estimates across the folds are then averaged and a pooled fluctuation step is performed to estimate the efficient influence function and derive pooled variance estimates.
+
+
+## Data and Notation
+
+### Our Data-Generating System 
+
+Consider $n$ observed units $O_1, \ldots, O_n$, where each unit $O_i$ is a random variable $(V \subset W, A, Y)$. The variable $W$ represents baseline covariates (such as age, sex, and education level), which are not considered to be effect modifiers. $V$ is a subset of $W$ that may be potential effect modifiers. $A$ is the exposure variable of interest (e.g., mixed pesticide biomarkers, metals, etc.), and $Y$ is the outcome of interest (e.g., disease status).
+
[…]
