### 2023-10-09 #5422 @blind-contours
I've added a new directory /doc that should have an html file that brings up the vignette. I'm going to make some edits to it now but it should be available for users.
