### 2023-10-19 #5422 @csoneson
@lucasmiranda42 - just wanted to check whether the additional information provided by @blind-contours allowed you to access the vignette?
