### 2023-10-21 #5422 @csoneson
@blind-contours I have gone through the submission and overall it looks good. Just a couple of points, mostly related to the bibliography - could you please:

- remove unused entries from the .bib file
- add a DOI for McCoy et al (2023)
- add `{}` as appropriate to make the representation of Mark van der Laan's name consistent across bibliography items
- add a space between `superNOVA` and package on line 100

In addition, the sentence starting on line 51 and ending on line 57 is very long, and I would suggest splitting it up to increase readability.

You can generate a new proof with `@editorialbot generate pdf`

Once this is done, please have a look at the next steps in the post below.
