### 2023-10-24 69bf8a3
Updates to JOSS paper.
