+++ inst/references.bib
@@ -1,3 +1,15 @@
+@article{SL_2008,
+title = {Super Learner},
+author = {van der Laan Mark, J. and C, Polley Eric and E., Hubbard Alan},
+year = {2007},
+journal = {Statistical Applications in Genetics and Molecular Biology},
+volume = {6},
+number = {1},
+pages = {1-23},
+url = {https://EconPapers.repec.org/RePEc:bpj:sagmbi:v:6:y:2007:i:1:n:25}
+}
+
+
 @misc{mccoy2023semiparametric,
       title={Semi-Parametric Identification and Estimation of Interaction and Effect Modification in Mixed Exposures using Stochastic Interventions},
       author={David B. McCoy and Alan E. Hubbard and Alejandro Schuler and Mark J. van der Laan},

[…]
