+++ paper/paper.bib
@@ -72,5 +72,6 @@
     year={2012},
     publisher={Wiley Online Library},
     url = {https://doi.org/10.1111/j.1541-0420.2011.01685.x},
-    doi = {10.1111/j.1541-0420
+    doi = {10.1111/j.1541-0420}
+    }
 
