### 2023-10-24 #5422 @blind-contours
I think I've made the suggested changes and also created the release version and upload to Zenodo correctly - if there are issues do please let me know.
