### 2023-10-25 #5422 @csoneson
Thanks @blind-contours - a couple of things remain:

- The capitalization of 'van der Laan' is still different for Diaz and van der Laan (2012).
- Please remove the 'https://doi.org/' from the DOI of McCoy et al (2023).
- I believe the invalid DOI for Diaz and van der Laan (2012) should be 10.1111/j.1541-0420.2011.01685.x
- The very long sentence (now lines 50-56, 'In the event of basis functions...') is still very long and hard to parse.
- Please make the author list in Zenodo the same as that of the paper
- I'm a bit confused by the version tags. The version in the DESCRIPTION file is 1.0, the version of the Zenodo archive you link to above is 1.0.0, and the latest release in the GitHub repo is 1.0.1. Could you please clarify?

Thanks!
