+++ paper/paper.bib
@@ -27,17 +27,17 @@
 
 @misc{mccoy2023semiparametric,
       title={Semi-Parametric Identification and Estimation of Interaction and Effect Modification in Mixed Exposures using Stochastic Interventions},
-      author={David B. McCoy and Alan E. Hubbard and Alejandro Schuler and {Van Der Laan}, Mark J.},
+      author={David B. McCoy and Alan E. Hubbard and Alejandro Schuler and {van der Laan}, Mark J.},
       year={2023},
       eprint={2305.01849},
       archivePrefix={arXiv},
       primaryClass={stat.ME},
-      doi = {https://doi.org/10.48550/arXiv.2305.01849}
+      doi = {10.48550/arXiv.2305.01849}
 }
 
 @article{Hubbard2016,
     abstract = {Consider one observes n i.i.d. copies of a random variable with a probability distribution that is known to be an element of a particular statistical model. In order to define our statistical target we partition the sample in V equal size sub-samples, and use this partitioning to define V splits in an estimation sample (one of the V subsamples) and corresponding complementary parameter-generating sample. For each of the V parameter-generating samples, we apply an algorithm that maps the sample to a statistical target parameter. We define our sample-split data adaptive statistical target parameter as the average of these V-sample specific target parameters. We present an estimator (and corresponding central limit theorem) of this type of data adaptive target parameter. This general methodology for generating data adaptive target parameters is demonstrated with a number of practical examples that highlight new opportunities for statistical learning from data. This new framework provides a rigorous statistical methodology for both exploratory and confirmatory analysis within the same data. Given that more research is becoming "data-driven", the theory developed within this paper provides a new impetus for a greater involvement of statistical inference into problems that are being increasingly addressed by clever, yet ad hoc pattern finding methods. To suggest such potential, and to verify the predictions of the theory, extensive simulation studies, along with a data analysis based on adaptively determined intervention rules are shown and give insight into how to structure such an approach. The results show that the data adaptive target parameter approach provides a general framework and resulting methodology for data-driven science.},
-    author = {Hubbard, Alan E. and Kherad-Pajouh, Sara and {Van Der Laan}, Mark J.},
+    author = {Hubbard, Alan E. and Kherad-Pajouh, Sara and {van der Laan}, Mark J.},
     doi = {10.1515/ijb-2015-0013},
     issn = {15574679},
     journal = {International Journal of Biostatistics},
@@ -52,7 +52,7 @@
 
 @article{Zheng2010,
[…]
