### 2023-11-01 #5422 @blind-contours
I believe this is the contributors section on Zenodo - I have all the authors listed as creators so I just removed that section and everything should match now. Let me know if you need anything else.
