### 2023-11-02 #5422 @csoneson
Sorry, I was unclear - the author lists match, but I see this in Zenodo:

```
McCoy, David; Alejandro, Shuler; Alan, Hubbard; Mark, van der Laan
```

I.e., the first names of your co-authors seem to be interpreted as surnames, and vice versa.
